% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{adjust.face.indices.to}
\alias{adjust.face.indices.to}
\title{Adjust integer matrix to target min value.}
\usage{
adjust.face.indices.to(faces, target_min_index = 1L)
}
\arguments{
\item{faces}{3xn integer matrix, the vertex indices of the faces}

\item{target_min_index}{integer, one of 1L or 0L. The target minimal value that the data should have afterwards.}
}
\value{
3xn integer matrix, the adjusted values
}
\description{
This takes a matrix of integers, and adjusts the values such that the minimal value is the 'target_min_index' value. It is used to adjust from 0-based to 1-based indices in meshes.
}
\note{
The current and the target min values must be 0 or 1.
}
\keyword{internal}
