% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R
\name{afni.tree}
\alias{afni.tree}
\title{Get tree of elements in AFNI dset file.}
\usage{
afni.tree(filepath)
}
\arguments{
\item{filepath}{character string, path to a file in AFNI dset format.}
}
\value{
a tree
}
\description{
Get tree of elements in AFNI dset file.
}
