% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{all.infinite.rows}
\alias{all.infinite.rows}
\title{Check which rows of a matrix consist of infinite values only.}
\usage{
\method{all}{infinite.rows}(mat)
}
\arguments{
\item{mat}{numeric matrix.}
}
\value{
logical vector with one entry per row.
}
\description{
Check which rows of a matrix consist of infinite values only.
}
\keyword{internal}
