% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_analyze.R
\name{analyze.write.data.internal}
\alias{analyze.write.data.internal}
\title{Write the voxel data of an ANALYZE 7.5 image to a connection.}
\usage{
analyze.write.data.internal(fh, analyzedata, analyzeheader)
}
\arguments{
\item{fh}{connection to write to.}

\item{analyzedata}{array of numeric data.}

\item{analyzeheader}{named list, a valid ANALYZE 7.5 header that describes the data.}
}
\value{
the data that was written, after conversion to the data type stated in the header.
}
\description{
Write the voxel data of an ANALYZE 7.5 image to a connection.
}
\keyword{internal}
