% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_analyze.R
\name{analyzeheader.for.data}
\alias{analyzeheader.for.data}
\title{Create ANALYZE 7.5 header suitable for given data.}
\usage{
analyzeheader.for.data(analyzedata, ...)
}
\arguments{
\item{analyzedata}{array of numeric (integer or double) data, can have up to 7 dimensions.}

\item{...}{extra parameters passed on to \code{\link{analyzeheader.template}}, ignored in this function.}
}
\value{
an ANALYZE 7.5 header (see \code{\link{analyzeheader.template}}) in which the \code{datatype}, \code{bitpix},
\code{dim}, \code{cal_min} and \code{cal_max} fields have been set to values suitable for the given data. Feel free to change
the other fields, e.g. the voxel sizes in \code{pix_dim} or the \code{orient} code.
}
\description{
Create ANALYZE 7.5 header suitable for given data.
}
\note{
ANALYZE 7.5 has much fewer data types than NIFTI: integers are stored as 32 bit signed integers and floating
point data as 32 bit floats, which is what this function selects (the same choice that
\code{\link{ni1header.for.data}} makes for NIFTI v1). To store 16 bit integers, which is the classic ANALYZE
data type, set \code{datatype} to \code{4L} and \code{bitpix} to \code{16L} in the returned header after calling this function.
}
\examples{
analyzeheader <- analyzeheader.for.data(array(1:24, dim = c(4, 3, 2)))
analyzeheader$datatype

}
