% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_annot.R
\name{annot.max.region.idx}
\alias{annot.max.region.idx}
\title{Get max region index of an fs.annot instance.}
\usage{
annot.max.region.idx(annot)
}
\arguments{
\item{annot}{fs.annot instance}
}
\value{
integer, the max region index. They typically start with 0 and are consecutive, but this is not enforced or checked in any way.
}
\description{
Get max region index of an fs.annot instance.
}
\note{
This is a helper function to be used with \code{annot.unique}, see the example there.
}
