% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvis_affine.R
\name{apply.affine.to.coords}
\alias{apply.affine.to.coords}
\title{Apply an affine to a set of coordinates.}
\usage{
apply.affine.to.coords(coords, affine)
}
\arguments{
\item{coords}{numeric matrix with 3 columns, the coordinates.}

\item{affine}{4x4 numeric matrix.}
}
\value{
numeric matrix with 3 columns, the transformed coordinates.
}
\description{
Apply an affine to a set of coordinates.
}
\keyword{internal}
