% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dti_tracks.R
\name{as.fs.tracts}
\alias{as.fs.tracts}
\title{Convert a collection of tracts to an fs.tracts instance.}
\usage{
as.fs.tracts(tracts, kind = "tck")
}
\arguments{
\item{tracts}{the input, see the description.}

\item{kind}{character string, 'tck' or 'trk', used for the returned instance.}
}
\value{
an \code{fs.tracts} instance.
}
\description{
Accepts an \code{fs.tracts} instance, a list of numeric matrices
with 3 columns, or a single such matrix (which is then treated as one
tract), and returns an \code{fs.tracts} instance. This is what the writers
use to accept several input types.
}
\examples{
tracts <- as.fs.tracts(list(matrix(c(0, 0, 0, 1, 1, 1), ncol = 3, byrow = TRUE)));
length(tracts);

}
