% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{as.list.fs.tracts}
\alias{as.list.fs.tracts}
\title{Convert an fs.tracts instance to a plain list of tracts.}
\usage{
\method{as.list}{fs.tracts}(x, ...)
}
\arguments{
\item{x}{an \code{fs.tracts} instance.}

\item{...}{ignored.}
}
\value{
list with one entry per tract.
}
\description{
Converts to the classic representation as a \code{list} with one
entry per tract. This is convenient for interacting with code that expects
a plain list, but note that it materializes one R object per tract and is
therefore much more expensive in both memory and time. Prefer \code{[[}
for accessing individual tracts.
}
\examples{
\dontrun{
tck <- read.dti.tck("brain.tck");
tracts_list <- as.list(tck$tracks);
}

}
