% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvis_affine.R
\name{axcodes2ornt}
\alias{axcodes2ornt}
\title{Convert axis codes to an orientation array.}
\usage{
axcodes2ornt(axcodes)
}
\arguments{
\item{axcodes}{character vector with one code per input axis, e.g.
\code{c('L','P','S')} for a file whose first voxel axis points left.}
}
\value{
numeric matrix with 3 rows and 2 columns, see
\code{\link{io.orientation}}.
}
\description{
Convert axis codes to an orientation array.
}
\keyword{internal}
