% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_mgh.R
\name{check.dtype.for.data}
\alias{check.dtype.for.data}
\title{Check whether the dtype is suitable for the data.}
\usage{
check.dtype.for.data(mridata, mri_dtype_code)
}
\arguments{
\item{mridata}{the data to check, a vector, matrix or array typically}

\item{mri_dtype_code}{integer, the MRI data type code. See \code{\link[freesurferformats]{translate.mri.dtype}}.}

\item{silent}{logical, whether to suppress warning messages if the dtype seems unsuitable}
}
\value{
logical, whether the dtype could be suitable. This is only a guess, as the checks are in no way complete.
}
\description{
This function provides an educated guess on whether the given dtype is suitable for the data. It is usually called for the site effect of printing warnings in case something seems off.
}
\keyword{internal}
