% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_writer.R
\name{check_data_and_settings_consistency}
\alias{check_data_and_settings_consistency}
\title{Warn about common errors in combining data and datatype.}
\usage{
check_data_and_settings_consistency(index, data, datatype, intent,
  force = FALSE)
}
\arguments{
\item{index}{positive integer, the dataarray index to report. Makes it easier for the user to find the broken one.}

\item{data}{vector or matrix, the data to write to the GIFTI file. Checked against the datatype.}

\item{datatype}{NIFTI datatype string, the datatype to use when writing to the GIFTI file. Checked against the data.}

\item{intent}{NIFTI intent string, checked independently. In no way do we check whether it makes sense for the data.}
}
\description{
Warn about common errors in combining data and datatype.
}
\note{
The checks in here are in no way exhaustive.
}
\keyword{internal}
