% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_parcels.R
\name{cifti.annot.list}
\alias{cifti.annot.list}
\title{Accept the input forms of a set of annotations.}
\usage{
cifti.annot.list(annots, structure = NULL)
}
\arguments{
\item{annots}{an \code{fs.annot} instance or a named list of them.}

\item{structure}{character string or \code{NULL}, the brain structure of a single
annotation.}
}
\value{
a named list of \code{fs.annot} instances, named by canonical brain structure name.
}
\description{
Accept the input forms of a set of annotations.
}
\keyword{internal}
