% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.array.with.new.first.dim}
\alias{cifti.array.with.new.first.dim}
\title{Create an array like the input, with a different first dimension.}
\usage{
cifti.array.with.new.first.dim(data, first_dim_size)
}
\arguments{
\item{data}{the array, its first dimension is the one that is replaced.}

\item{first_dim_size}{integer, the size of the new first dimension.}
}
\value{
an array of \code{NA} of the data type of \code{data}, with the new first
dimension and all other dimensions of \code{data}.
}
\description{
Create an array like the input, with a different first dimension.
}
\keyword{internal}
