% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_axes.R
\name{cifti.axis.from.template}
\alias{cifti.axis.from.template}
\title{Create a CIFTI-2 axis from a template file.}
\usage{
cifti.axis.from.template(x, dim = NULL)
}
\arguments{
\item{x}{character string (the path of a CIFTI-2 file), an \code{fs.cifti} object (see
\code{\link{read.cifti.header}}) or an \code{fs.cifti.data} object (see
\code{\link{read.cifti}}).}

\item{dim}{integer or \code{NULL}, the matrix dimension to get the axis for. If \code{NULL} (the
default), the axes of all matrix dimensions are returned as a list.}
}
\value{
the axis of the requested dimension (a named list, see
\code{\link{cifti.axis.brain.models}}), or a list of axes (one per dimension; note
that a file whose dimensions share a single XML mapping returns the same axis for
both, as two entries).
}
\description{
Read the axes of an existing CIFTI-2 file, so that a new file can be
written with the same mapping. This is the recommended way to write a file for real
data: the mapping of an HCP grayordinates file (which vertices are in the file, and
which volume voxels) can not be invented, it has to come from the file the data was
derived from.
}
\examples{
cifti_file <- system.file("extdata", "cifti", "tiny.dscalar.nii", package = "freesurferformats")
axes <- cifti.axis.from.template(cifti_file)
length(axes)
axes[[2L]]$type

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels]{cifti.axis.parcels()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.axis.series]{cifti.axis.series()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.file.type.for.axes]{cifti.file.type.for.axes()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=cifti.volume]{cifti.volume()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
