% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.axis.size}
\alias{cifti.axis.size}
\title{The number of matrix entries an axis covers.}
\usage{
cifti.axis.size(axis)
}
\arguments{
\item{axis}{an axis, see \code{\link{cifti.header.from.axes}}.}
}
\value{
integer, the size of that matrix dimension.
}
\description{
The number of matrix entries an axis covers.
}
\keyword{internal}
