% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.axis.surface.sizes}
\alias{cifti.axis.surface.sizes}
\title{Get the surface sizes of an axis as a named vector.}
\usage{
cifti.axis.surface.sizes(axis)
}
\arguments{
\item{axis}{an axis of type 'CIFTI_INDEX_TYPE_BRAIN_MODELS'.}
}
\value{
named integer vector or \code{NULL}, the number of vertices per structure, see
\code{\link{cifti.axis.brain.models}}.
}
\description{
Get the surface sizes of an axis as a named vector.
}
\keyword{internal}
