% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.axis.surfaces}
\alias{cifti.axis.surfaces}
\title{Create the Surface elements of an axis.}
\usage{
cifti.axis.surfaces(surfaces)
}
\arguments{
\item{surfaces}{named integer vector or \code{NULL}, the number of vertices per brain
structure, see \code{\link{cifti.axis.brain.models}}.}
}
\value{
a list of lists with the entries 'brain_structure' and
'surface_number_of_vertices', or \code{NULL}.
}
\description{
Create the Surface elements of an axis.
}
\keyword{internal}
