% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.brainordinate.dim}
\alias{cifti.brainordinate.dim}
\title{Get the brainordinate dimension of a CIFTI-2 file.}
\usage{
cifti.brainordinate.dim(cii)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}
}
\value{
integer, the dimension. Stops with a descriptive error if there is no
brainordinate dimension or if there are two.
}
\description{
Find the matrix dimension of a CIFTI-2 file that holds the brain
structures. Most files have exactly one (dimension 1 for a \code{.dscalar},
\code{.dtseries} or \code{.dlabel}, dimension 0 for the special case of a
\code{.dpconn}-style file), while connectome files like a \code{.dconn} have brain
models in both dimensions.
}
\keyword{internal}
