% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.check.axes}
\alias{cifti.check.axes}
\title{Check a list of axes and name the dimensions.}
\usage{
cifti.check.axes(axes)
}
\arguments{
\item{axes}{list of axes, see \code{\link{cifti.header.from.axes}}.}
}
\value{
the list, with the axes named by their matrix dimension.
}
\description{
Check a list of axes and name the dimensions.
}
\keyword{internal}
