% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.check.data.for.axes}
\alias{cifti.check.data.for.axes}
\title{Check and prepare the data for a set of axes.}
\usage{
cifti.check.data.for.axes(data, axes)
}
\arguments{
\item{data}{the data, see \code{\link{write.cifti}}.}

\item{axes}{list of axes, see \code{\link{cifti.header.from.axes}}.}
}
\value{
the data as an array with the dimensions of the axes.
}
\description{
Check and prepare the data for a set of axes.
}
\keyword{internal}
