% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.check.index.list}
\alias{cifti.check.index.list}
\title{Check a vector of 0-based indices.}
\usage{
cifti.check.index.list(indices, what, max_index = NULL)
}
\arguments{
\item{indices}{integer vector, the indices.}

\item{what}{character string, a description of the indices, used in error messages.}

\item{max_index}{integer or \code{NULL}, the largest allowed index.}
}
\value{
the indices as an integer vector.
}
\description{
Check a vector of 0-based indices.
}
\keyword{internal}
