% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.check.index.selection}
\alias{cifti.check.index.selection}
\title{Check a matrix index selection.}
\usage{
cifti.check.index.selection(selection, max_index, what, filepath)
}
\arguments{
\item{selection}{integer vector or \code{NULL}, the indices to check.}

\item{max_index}{integer, the size of the matrix dimension.}

\item{what}{character string, the name of the selection ('rows' or 'columns'),
used in the error message.}

\item{filepath}{character string, the path of the file, used in the error message.}
}
\value{
the selection as an integer vector, or \code{NULL}.
}
\description{
Check a matrix index selection.
}
\keyword{internal}
