% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.check.label.table}
\alias{cifti.check.label.table}
\title{Check a label table for writing.}
\usage{
cifti.check.label.table(label_table, map_idx)
}
\arguments{
\item{label_table}{a data.frame, see \code{\link{cifti.axis.labels}}, or \code{NULL}.}

\item{map_idx}{integer, the number of the map, used in the error message.}
}
\value{
the label table, or \code{NULL}.
}
\description{
Check a label table for writing.
}
\keyword{internal}
