% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.check.object}
\alias{cifti.check.object}
\title{Check that an object is an fs.cifti instance.}
\usage{
cifti.check.object(cii)
}
\arguments{
\item{cii}{the object to check.}
}
\value{
the object, invisibly.
}
\description{
Check that an object is an fs.cifti instance.
}
\keyword{internal}
