% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_connectome.R
\name{cifti.connectome.axes}
\alias{cifti.connectome.axes}
\title{Determine the two axes of a connectome file.}
\usage{
cifti.connectome.axes(axes, template_cii = NULL)
}
\arguments{
\item{axes}{list of one or two axes or \code{NULL}, see
\code{\link{write.fs.connectome.cifti}}.}

\item{template_cii}{an \code{fs.cifti} object or \code{NULL}, see
\code{\link{write.fs.connectome.cifti}}.}
}
\value{
a list of two axes, named by matrix dimension.
}
\description{
Determine the two axes of a connectome file.
}
\keyword{internal}
