% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_connectome.R
\name{cifti.data.for.connectome}
\alias{cifti.data.for.connectome}
\title{Accept the input forms of a CIFTI-2 connectome.}
\usage{
cifti.data.for.connectome(filepath, rows = NULL, columns = NULL)
}
\arguments{
\item{filepath}{character string, \code{fs.cifti} or \code{fs.cifti.data}, see
\code{\link{read.fs.connectome.cifti}}.}

\item{rows}{integer vector or \code{NULL}, the indices of matrix dimension 0 to read.}

\item{columns}{integer vector or \code{NULL}, the indices of matrix dimension 1 to read.}
}
\value{
an \code{fs.cifti.data} object, see \code{\link{read.cifti}}.
}
\description{
Accept the input forms of a CIFTI-2 connectome.
}
\keyword{internal}
