% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.data.for.grayordinates}
\alias{cifti.data.for.grayordinates}
\title{Select the data values of the grayordinates of a file.}
\usage{
cifti.data.for.grayordinates(data, structures, index_table, what)
}
\arguments{
\item{data}{the data, see \code{\link{write.fs.morph.cifti}}.}

\item{structures}{character vector, the canonical structure names, in file order.}

\item{index_table}{the brainordinate table, see \code{\link{cifti.grayordinates}}.}

\item{what}{character string, a description of the data, used in error messages.}
}
\value{
a matrix with one row per grayordinate (in file order) and one column per map.
}
\description{
Select the data values of the grayordinates of a file.
}
\keyword{internal}
