% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.data.object}
\alias{cifti.data.object}
\title{Accept the input forms of a CIFTI-2 data object.}
\usage{
cifti.data.object(x)
}
\arguments{
\item{x}{an \code{fs.cifti.data} object, an \code{fs.cifti} metadata object or the path
of a CIFTI-2 file.}
}
\value{
an \code{fs.cifti.data} object, see \code{\link{read.cifti}}.
}
\description{
Accept the input forms of a CIFTI-2 data object.
}
\keyword{internal}
