% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.data.per.structure}
\alias{cifti.data.per.structure}
\title{Get the per-structure data of one structure.}
\usage{
cifti.data.per.structure(data, structure_name, what)
}
\arguments{
\item{data}{the data, see \code{\link{write.fs.morph.cifti}}.}

\item{structure_name}{character string, the canonical structure name.}

\item{what}{character string, a description of the data, used in error messages.}
}
\value{
a matrix with one row per vertex of the complete surface and one column per map.
}
\description{
Get the per-structure data of one structure.
}
\keyword{internal}
