% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.data.structures}
\alias{cifti.data.structures}
\title{Determine the structures of per-vertex data.}
\usage{
cifti.data.structures(data, structure = NULL)
}
\arguments{
\item{data}{the data, see \code{\link{write.fs.morph.cifti}}.}

\item{structure}{character string or \code{NULL}, the structure for the non-list case.}
}
\value{
character vector of canonical structure names, or \code{NULL} if the data are not a
named list and no structure was given.
}
\description{
Determine the structures of per-vertex data.
}
\keyword{internal}
