% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.data.surface.size}
\alias{cifti.data.surface.size}
\title{The number of vertices of the surface of one structure.}
\usage{
cifti.data.surface.size(data, structure_name)
}
\arguments{
\item{data}{the data, see \code{\link{write.fs.morph.cifti}}.}

\item{structure_name}{character string, the canonical structure name.}
}
\value{
integer, the number of vertices.
}
\description{
The number of vertices of the surface of one structure.
}
\keyword{internal}
