% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.dim.to.front}
\alias{cifti.dim.to.front}
\title{Move one dimension of an array to the front.}
\usage{
cifti.dim.to.front(data, dim)
}
\arguments{
\item{data}{the array.}

\item{dim}{integer, the dimension to move, counted from 0 (CIFTI style).}
}
\value{
the array with the requested dimension first, and all other dimensions
in their original order.
}
\description{
Move one dimension of an array to the front.
}
\keyword{internal}
