% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.file.looks.like.cifti1}
\alias{cifti.file.looks.like.cifti1}
\title{Check whether a file looks like a CIFTI-1 file.}
\usage{
cifti.file.looks.like.cifti1(filepath)
}
\arguments{
\item{filepath}{character string, the path to a CIFTI-2 file (usually one of
\code{.dscalar.nii}, \code{.dtseries.nii}, \code{.dlabel.nii}, \code{.dconn.nii}, \code{.pscalar.nii},
\code{.ptseries.nii}, \code{.pconn.nii}, \code{.dpconn.nii} or \code{.pdconn.nii}). Note that
this is not a NIFTI file, despite the \code{.nii} part. Gzipped CIFTI files are
not supported, because the CIFTI-2 format forbids compression.}
}
\value{
logical, whether the file looks like a CIFTI-1 file.
}
\description{
This is a heuristic used to produce a helpful error message when
a CIFTI-1 file is passed to the reader: CIFTI-1 files are NIFTI-1 files (not
NIFTI-2), and this function searches the first bytes of the file for the
string 'CIFTI'. It is not used for any other purpose.
}
\keyword{internal}
