% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_connectome.R
\name{cifti.file.type.for.extension}
\alias{cifti.file.type.for.extension}
\title{The CIFTI-2 file type a file name names.}
\usage{
cifti.file.type.for.extension(filepath)
}
\arguments{
\item{filepath}{character string, a file name.}
}
\value{
character string, the standard file name extension the name ends with (e.g.
'.pscalar.nii'), or the empty string if it does not end with one of them.
}
\description{
The CIFTI-2 file type a file name names.
}
\keyword{internal}
