% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_fs.R
\name{cifti.grayordinates.for.axes}
\alias{cifti.grayordinates.for.axes}
\title{Build the brainordinate table of an axis.}
\usage{
cifti.grayordinates.for.axes(axis)
}
\arguments{
\item{axis}{an axis of type 'CIFTI_INDEX_TYPE_BRAIN_MODELS'.}
}
\value{
a data.frame, see \code{\link{cifti.grayordinates}}.
}
\description{
Build the brainordinate table of an axis.
}
\keyword{internal}
