% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.grayordinates.for.model}
\alias{cifti.grayordinates.for.model}
\title{Expand one brain model entry into a brainordinate table.}
\usage{
cifti.grayordinates.for.model(map, model)
}
\arguments{
\item{map}{the indices map the brain model belongs to, see
\code{\link{cifti.map.for.dim}}.}

\item{model}{one brain model entry of the map.}
}
\value{
a data.frame, see \code{\link{cifti.grayordinates}}.
}
\description{
Expand one brain model entry into a brainordinate table.
}
\keyword{internal}
