% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.header.of}
\alias{cifti.header.of}
\title{Accept the input forms of a CIFTI-2 header.}
\usage{
cifti.header.of(x)
}
\arguments{
\item{x}{character string (a file path), an \code{fs.cifti} or an \code{fs.cifti.data} object.}
}
\value{
an \code{fs.cifti} object, see \code{\link{read.cifti.header}}.
}
\description{
Accept the input forms of a CIFTI-2 header.
}
\keyword{internal}
