% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_connectome.R
\name{cifti.index.type.short}
\alias{cifti.index.type.short}
\title{The short name of a CIFTI-2 matrix index type.}
\usage{
cifti.index.type.short(type)
}
\arguments{
\item{type}{character string, a CIFTI-2 matrix index type, see
\code{\link{cifti.header.from.axes}}.}
}
\value{
character string, the type without the prefix, e.g. 'BRAIN_MODELS'.
}
\description{
The short name of a CIFTI-2 matrix index type.
}
\keyword{internal}
