% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.index.types}
\alias{cifti.index.types}
\title{The CIFTI-2 index types (mapping types).}
\usage{
cifti.index.types()
}
\description{
The CIFTI-2 index types (mapping types).
}
\keyword{internal}
