% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_connectome.R
\name{cifti.is.connectome.axis}
\alias{cifti.is.connectome.axis}
\title{Check whether an axis describes a connectome dimension.}
\usage{
cifti.is.connectome.axis(axis)
}
\arguments{
\item{axis}{an axis, see \code{\link{cifti.header.from.axes}}.}
}
\value{
logical, whether the axis holds brainordinates or parcels.
}
\description{
Check whether an axis describes a connectome dimension.
}
\keyword{internal}
