% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti.R
\name{cifti.map.name}
\alias{cifti.map.name}
\title{Get the name of one named map of a CIFTI-2 file.}
\usage{
cifti.map.name(cii, dim, map = 1L)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension that holds the named maps.}

\item{map}{integer, the number of the map (1-based).}
}
\value{
character string, the map name, or \code{NULL} if the map is unnamed.
}
\description{
Get the name of one named map of a CIFTI-2 file.
}
\keyword{internal}
