% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.map.type.description}
\alias{cifti.map.type.description}
\title{Describe a CIFTI-2 index type for humans.}
\usage{
cifti.map.type.description(type)
}
\arguments{
\item{type}{character string, a CIFTI-2 index type, see \code{\link{cifti.index.types}}.}
}
\value{
character string, a short description.
}
\description{
Describe a CIFTI-2 index type for humans.
}
\keyword{internal}
