% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.matrix.dim.sizes}
\alias{cifti.matrix.dim.sizes}
\title{Determine the sizes of the CIFTI matrix dimensions.}
\usage{
cifti.matrix.dim.sizes(niiheader, filepath = "")
}
\arguments{
\item{niiheader}{the NIFTI-2 header of the file, required to determine the
sizes of the matrix dimensions.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
integer vector, the sizes of the matrix dimensions.
}
\description{
The CIFTI-2 matrix dimensions 0, 1, ... are stored in the
\code{dim} field of the NIFTI-2 header, starting at its 6th entry (R index
vectors are 1-based, so matrix dimension 0 is entry 6). The dimension order
in the XML refers to matrix dimensions, so a \code{MatrixIndicesMap} with
\code{AppliesToMatrixDimension="0"} describes the 6th entry of \code{dim}.
}
\keyword{internal}
