% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.merge.axes}
\alias{cifti.merge.axes}
\title{Merge explicitly given axes with the axes of a template.}
\usage{
cifti.merge.axes(axes, template_cii)
}
\arguments{
\item{axes}{list of axes or \code{NULL}, see \code{\link{write.cifti}}.}

\item{template_cii}{an \code{fs.cifti} object or \code{NULL}.}
}
\value{
the merged list of axes, named by matrix dimension.
}
\description{
Merge explicitly given axes with the axes of a template.
}
\keyword{internal}
