% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti.R
\name{cifti.other.dim}
\alias{cifti.other.dim}
\title{Get the other matrix dimension of a 2-dimensional CIFTI-2 matrix.}
\usage{
cifti.other.dim(cii, dim)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension to exclude.}
}
\value{
integer, the other matrix dimension. Stops for files with more than two
matrix dimensions.
}
\description{
Get the other matrix dimension of a 2-dimensional CIFTI-2 matrix.
}
\keyword{internal}
