% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_connectome.R
\name{cifti.parcels.axis.of}
\alias{cifti.parcels.axis.of}
\title{Get the parcels axis of a template file.}
\usage{
cifti.parcels.axis.of(template_cii)
}
\arguments{
\item{template_cii}{an \code{fs.cifti} object, see \code{\link{cifti.header.of}}.}
}
\value{
the parcels axis of the file, see \code{\link{cifti.axis.parcels}}.
}
\description{
Get the parcels axis of a template file.
}
\keyword{internal}
