% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.child.int.vector}
\alias{cifti.parse.child.int.vector}
\title{Parse a required or optional child element holding integer values.}
\usage{
cifti.parse.child.int.vector(node, child_name, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the parent element.}

\item{child_name}{character string, the name of the child element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
integer vector, or \code{NULL} if the child element is missing or empty.
}
\description{
Parse a required or optional child element holding integer values.
}
\keyword{internal}
