% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.indices.map}
\alias{cifti.parse.indices.map}
\title{Parse one MatrixIndicesMap element.}
\usage{
cifti.parse.indices.map(node, dim_sizes, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the \code{MatrixIndicesMap} element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
a named list, see \code{\link{read.cifti.header}}.
}
\description{
Parse one MatrixIndicesMap element.
}
\keyword{internal}
