% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.label.table}
\alias{cifti.parse.label.table}
\title{Parse a CIFTI LabelTable element.}
\usage{
cifti.parse.label.table(node)
}
\arguments{
\item{node}{an xml2 node, the \code{LabelTable} element, or an \code{xml_missing} node.}
}
\value{
a data.frame with the columns 'key', 'red', 'green', 'blue', 'alpha',
'label', 'x', 'y' and 'z', or \code{NULL} if the node is missing.
}
\description{
Parse a CIFTI LabelTable element.
}
\keyword{internal}
