% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.metadata}
\alias{cifti.parse.metadata}
\title{Parse the MetaData element of a CIFTI XML node.}
\usage{
cifti.parse.metadata(node)
}
\arguments{
\item{node}{an xml2 node that may contain a \code{MetaData} child element.}
}
\value{
a named list of character strings, the metadata entries in file order.
Empty list if the node has no metadata.
}
\description{
Parse the MetaData element of a CIFTI XML node.
}
\keyword{internal}
