% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.named.maps}
\alias{cifti.parse.named.maps}
\title{Parse the NamedMap elements of a MatrixIndicesMap element.}
\usage{
cifti.parse.named.maps(node, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the \code{MatrixIndicesMap} element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
a list of lists with the entries 'name', 'metadata' and 'labels', or
\code{NULL} if there are no NamedMap elements.
}
\description{
Parse the NamedMap elements of a MatrixIndicesMap element.
}
\keyword{internal}
