% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.numeric.vector}
\alias{cifti.parse.numeric.vector}
\title{Parse a whitespace-separated list of numbers from XML text content.}
\usage{
cifti.parse.numeric.vector(text, what = "an element", context = "")
}
\arguments{
\item{text}{character string, the text content of an XML element.}

\item{what}{character string, the element/attribute name, used in error messages.}

\item{context}{character string, additional context for error messages.}
}
\value{
numeric vector. Empty numeric vector for empty input.
}
\description{
Parse a whitespace-separated list of numbers from XML text content.
}
\keyword{internal}
