% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.parcels}
\alias{cifti.parse.parcels}
\title{Parse the Parcel elements of a MatrixIndicesMap element.}
\usage{
cifti.parse.parcels(node, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the \code{MatrixIndicesMap} element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
a list of lists, or \code{NULL} if there are no Parcel elements.
}
\description{
Parse the Parcel elements of a MatrixIndicesMap element.
}
\keyword{internal}
