% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.surfaces}
\alias{cifti.parse.surfaces}
\title{Parse the Surface elements of a MatrixIndicesMap element.}
\usage{
cifti.parse.surfaces(node, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the \code{MatrixIndicesMap} element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
a list of lists with the entries 'brain_structure' and
'surface_number_of_vertices', or \code{NULL} if there are no Surface elements.
}
\description{
Parse the Surface elements of a MatrixIndicesMap element.
}
\keyword{internal}
