% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.volumes}
\alias{cifti.parse.volumes}
\title{Parse the Volume elements of a MatrixIndicesMap element.}
\usage{
cifti.parse.volumes(node, filepath = "")
}
\arguments{
\item{node}{an xml2 node, the \code{MatrixIndicesMap} element.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
a list of lists with the entries 'dimensions', 'meter_exponent' and
'transformation_matrix', or \code{NULL} if there are no Volume elements.
}
\description{
Parse the Volume elements of a MatrixIndicesMap element.
}
\keyword{internal}
