% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parse.xml}
\alias{cifti.parse.xml}
\title{Parse the CIFTI XML metadata.}
\usage{
cifti.parse.xml(xml_text, filepath = "", niiheader = NULL)
}
\arguments{
\item{xml_text}{character string, the XML document.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}

\item{niiheader}{the NIFTI-2 header of the file, required to determine the
sizes of the matrix dimensions.}
}
\value{
an \code{fs.cifti} object, see \code{\link{read.cifti.header}}.
}
\description{
Parse the CIFTI XML metadata.
}
\keyword{internal}
