% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.read.values}
\alias{cifti.read.values}
\title{Read raw values from the data section of a CIFTI-2 file.}
\usage{
cifti.read.values(cii, num_values, skip_values = 0L, fh = NULL)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{num_values}{integer, the number of values to read.}

\item{skip_values}{integer, the number of values to skip first.}

\item{fh}{optional connection to the file, positioned at the start of the data
section if \code{skip_values} is 0. Will be opened (and closed) if left at \code{NULL}.}
}
\value{
vector of values, see \code{\link{read.nifti.values}}.
}
\description{
Read raw values from the data section of a CIFTI-2 file.
}
\keyword{internal}
