% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti.R
\name{cifti.structure.data.dim}
\alias{cifti.structure.data.dim}
\title{Get the matrix dimension a brain structure lives in.}
\usage{
cifti.structure.data.dim(cii, structure)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{structure}{character string, the canonical name of the brain structure.}
}
\value{
integer, the matrix dimension.
}
\description{
The counterpart of \code{\link{cifti.other.dim}} for a structure
name: look up the dimension of the file that contains brain models for the
given structure. Needed by the user-facing readers, which get a structure name
from the user and must decide which dimension of the data matrix it refers to.
}
\keyword{internal}
