% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.structure.data.one}
\alias{cifti.structure.data.one}
\title{Extract the data of one brain structure (internal).}
\usage{
cifti.structure.data.one(cii, map, brainordinates, data, structure)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{map}{the indices map of the brainordinate dimension, see
\code{\link{cifti.map.for.dim}}.}

\item{brainordinates}{the brainordinate table, see
\code{\link{cifti.grayordinates}}.}

\item{data}{the data array with the brainordinate dimension first, see
\code{\link{cifti.dim.to.front}}.}

\item{structure}{character string, the canonical name of the structure.}
}
\value{
a named list, see \code{\link{cifti.structure.data}}.
}
\description{
Extract the data of one brain structure (internal).
}
\keyword{internal}
