% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.subset.dim}
\alias{cifti.subset.dim}
\title{Subset one dimension of an array.}
\usage{
cifti.subset.dim(data, dim, selection)
}
\arguments{
\item{data}{the array to subset.}

\item{dim}{integer, the dimension to subset.}

\item{selection}{integer vector or \code{NULL}, the indices to keep.}
}
\value{
the subset, with the dimensions of \code{data} preserved.
}
\description{
Subset one dimension of an array.
}
\keyword{internal}
