% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.validate.dims.coverage}
\alias{cifti.validate.dims.coverage}
\title{Check that all matrix dimensions are described exactly once.}
\usage{
cifti.validate.dims.coverage(indices_maps, dim_sizes, filepath = "")
}
\arguments{
\item{indices_maps}{list of parsed indices maps.}

\item{filepath}{character string, the path of the file the XML was read from.
Only used in error messages.}
}
\value{
\code{NULL}, invisibly. Stops if a dimension is described twice or not at
all.
}
\description{
Check that all matrix dimensions are described exactly once.
}
\keyword{internal}
