% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_axes.R
\name{cifti.volume}
\alias{cifti.volume}
\title{Create a CIFTI-2 axis for a volume.}
\usage{
cifti.volume(dimensions, transformation_matrix, meter_exponent = -3L)
}
\arguments{
\item{dimensions}{integer vector of length 3, the dimensions of the voxel grid.}

\item{transformation_matrix}{4x4 numeric matrix, the transformation from (0-based)
voxel indices to coordinates. There is no default: a CIFTI-2 file does not store the
voxel size or the position of a volume anywhere else, so a wrong or invented matrix
silently puts the voxels of a structure in the wrong place. Pass the matrix that the
template file has (see \code{\link{cifti.axis.from.template}}) if you do not know it.}

\item{meter_exponent}{integer, the exponent of the unit of the coordinates.}
}
\value{
a named list with the entries 'dimensions', 'meter_exponent' and
'transformation_matrix', to be passed to \code{\link{cifti.axis.brain.models}} or
\code{\link{cifti.axis.parcels}}.
}
\description{
A volume axis describes the voxel grid that the volume brain models of a
CIFTI-2 file refer to: its dimensions and the 4x4 transformation matrix that maps the
(0-based) IJK voxel indices to coordinates. The matrix is stored row by row in the
file, and the coordinates it produces are in units of \code{10^meter_exponent} (which is
-3, i.e. millimeters, for the files that Connectome Workbench writes).
}
\examples{
volume <- cifti.volume(c(4L, 4L, 4L), diag(c(2, 2, 2, 1)))
volume$dimensions

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.from.template]{cifti.axis.from.template()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels]{cifti.axis.parcels()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.axis.series]{cifti.axis.series()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.file.type.for.axes]{cifti.file.type.for.axes()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
