% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.xml.add.indices.map}
\alias{cifti.xml.add.indices.map}
\title{Add a MatrixIndicesMap element to the Matrix element.}
\usage{
cifti.xml.add.indices.map(matrix_node, axis, dims)
}
\arguments{
\item{matrix_node}{an xml2 node, the \code{Matrix} element.}

\item{axis}{the axis to write, see \code{\link{cifti.header.from.axes}}.}

\item{dims}{integer vector, the matrix dimensions the element applies to.}
}
\value{
\code{NULL}, invisibly. The element is added as a child of \code{matrix_node}.
}
\description{
Add a MatrixIndicesMap element to the Matrix element.
}
\keyword{internal}
