% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.xml.add.metadata}
\alias{cifti.xml.add.metadata}
\title{Add a MetaData element to an XML node.}
\usage{
cifti.xml.add.metadata(node, metadata)
}
\arguments{
\item{node}{an xml2 node, the parent element.}

\item{metadata}{named character vector, named list or \code{NULL}, the name/value pairs.}
}
\value{
\code{NULL}, invisibly. The element is added as a child of \code{node}.
}
\description{
Add a MetaData element to an XML node.
}
\keyword{internal}
