% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.xml.int.vector}
\alias{cifti.xml.int.vector}
\title{Format an integer vector for an XML text node.}
\usage{
cifti.xml.int.vector(x)
}
\arguments{
\item{x}{integer vector.}
}
\value{
character string, the values separated by single spaces.
}
\description{
Format an integer vector for an XML text node.
}
\keyword{internal}
