% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.xml.num.vector}
\alias{cifti.xml.num.vector}
\title{Format a numeric vector for an XML text node.}
\usage{
cifti.xml.num.vector(x, decimals = 10L)
}
\arguments{
\item{x}{numeric vector.}

\item{decimals}{integer, the number of decimal places.}
}
\value{
character string, the values separated by single spaces.
}
\description{
Format a numeric vector for an XML text node.
}
\keyword{internal}
