% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.xml.version}
\alias{cifti.xml.version}
\title{The CIFTI version this package writes.}
\usage{
cifti.xml.version()
}
\value{
character string, '2'.
}
\description{
The CIFTI version this package writes.
}
\keyword{internal}
