% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{CIFTI_EXTENSION_CODE}
\alias{CIFTI_EXTENSION_CODE}
\title{The NIFTI v2 extension code that holds the CIFTI-2 XML metadata.}
\usage{
CIFTI_EXTENSION_CODE
}
\description{
The CIFTI-2 XML document is stored in a NIFTI v2 header extension
with this code.
}
\keyword{internal}
