% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{coord.bbox}
\alias{coord.bbox}
\title{Compute the bounding box of a set of coordinates.}
\usage{
coord.bbox(coords)
}
\arguments{
\item{coords}{numeric matrix with 3 columns.}
}
\value{
numeric vector of length 6,
\code{c(xmin, xmax, ymin, ymax, zmin, zmax)}, or NULL if there are no rows.
}
\description{
Compute the bounding box of a set of coordinates.
}
\keyword{internal}
