% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{coord.to.key}
\alias{coord.to.key}
\title{Turn coordinate vector into string.}
\usage{
coord.to.key(coord, digits = 6L)
}
\arguments{
\item{coord}{double vector of length 3, the xyz coord}

\item{digits}{integer, the number of digits (after the decimal separator) to use}
}
\value{
character string
}
\description{
Turn coordinate vector into string.
}
\keyword{internal}
