% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{detect.dti.tract.format}
\alias{detect.dti.tract.format}
\title{Detect the format of a DTI tract file.}
\usage{
detect.dti.tract.format(filepath)
}
\arguments{
\item{filepath}{character string, path to the file.}
}
\value{
character string, one of 'tck', 'tsf' or 'trk'.
}
\description{
Detect the format of a DTI tract file.
}
\keyword{internal}
