% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dwi_gradients.R
\name{.name.gradient.table.columns}
\alias{.name.gradient.table.columns}
\title{Name the columns of an MRtrix gradient table.}
\usage{
.name.gradient.table.columns(gt)
}
\arguments{
\item{gt}{numeric matrix with 4 columns, one row per volume.}
}
\value{
the matrix with columns named 'x', 'y', 'z' and 'b'.
}
\description{
Name the columns of an MRtrix gradient table.
}
\keyword{internal}
