% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_dist.R
\name{euclidian.dist}
\alias{euclidian.dist}
\title{Compute Euclidian distance.}
\usage{
euclidian.dist(x1, x2)
}
\arguments{
\item{x1}{numerical vector, coords of first point}

\item{x2}{numerical vector, coords of second point}
}
\value{
the Euclidian distance between x1 and x2.
}
\description{
Compute Euclidian distance.
}
\keyword{internal}
