% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{faces.quad.to.tris}
\alias{faces.quad.to.tris}
\title{Convert quadrangular faces or polygons to triangular ones.}
\usage{
faces.quad.to.tris(quad_faces)
}
\arguments{
\item{quad_faces}{nx4 integer matrix, the indices of the vertices making up the \emph{n} quad faces.}
}
\value{
\emph{2nx3} integer matrix, the indices of the vertices making up the \emph{2n} tris faces.
}
\description{
Convert quadrangular faces or polygons to triangular ones.
}
\note{
This function does no fancy remeshing, it simply splits each quad into two triangles.
}
\seealso{
Other mesh functions:
\code{\link[=read.fs.surface]{read.fs.surface()}},
\code{\link[=read.fs.surface.asc]{read.fs.surface.asc()}},
\code{\link[=read.fs.surface.bvsrf]{read.fs.surface.bvsrf()}},
\code{\link[=read.fs.surface.geo]{read.fs.surface.geo()}},
\code{\link[=read.fs.surface.gii]{read.fs.surface.gii()}},
\code{\link[=read.fs.surface.ico]{read.fs.surface.ico()}},
\code{\link[=read.fs.surface.obj]{read.fs.surface.obj()}},
\code{\link[=read.fs.surface.off]{read.fs.surface.off()}},
\code{\link[=read.fs.surface.ply]{read.fs.surface.ply()}},
\code{\link[=read.fs.surface.vtk]{read.fs.surface.vtk()}},
\code{\link[=read.mesh.brainvoyager]{read.mesh.brainvoyager()}},
\code{\link[=read_nisurface]{read_nisurface()}},
\code{\link[=read_nisurfacefile]{read_nisurfacefile()}},
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.asc]{write.fs.surface.asc()}},
\code{\link[=write.fs.surface.byu]{write.fs.surface.byu()}},
\code{\link[=write.fs.surface.gii]{write.fs.surface.gii()}},
\code{\link[=write.fs.surface.mz3]{write.fs.surface.mz3()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh functions}
