% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{faces.tris.to.quad}
\alias{faces.tris.to.quad}
\title{Convert tris faces to quad faces by simple merging.}
\usage{
faces.tris.to.quad(tris_faces)
}
\arguments{
\item{tris_faces}{\emph{nx3} integer matrix, the indices of the vertices making up the \emph{n} tris faces.}
}
\value{
n/2x4 integer matrix, the indices of the vertices making up the \emph{n} quad faces.
}
\description{
This is experimental. Note that it can only work if the number of 'tris_faces' is even, as two consecutive tris-faces will be merged into one quad face. We could set the index to NA in that case, but I do not know how FreeSurfer handles this, so we do not guess.
}
\note{
This function does not implement proper remeshing of tri-meshes to quad-meshes. Use a proper mesh library if you need that.
}
