% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti1.R
\name{fileopen.gz.or.not}
\alias{fileopen.gz.or.not}
\title{Get connection to a binary file, gz or not.}
\usage{
fileopen.gz.or.not(filepath)
}
\arguments{
\item{filepath}{path to the binary file.}
}
\description{
Get connection to a binary file, gz or not.
}
\keyword{internal}
