% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_analyze.R
\name{fileopen.write.gz.or.not}
\alias{fileopen.write.gz.or.not}
\title{Open a connection for writing, with gzip support based on the file name.}
\usage{
fileopen.write.gz.or.not(filepath)
}
\arguments{
\item{filepath}{character string, the path of the file to write.}
}
\value{
a connection, either a \code{gzfile} or a plain file connection.
}
\description{
Open a connection for writing, with gzip support based on the file name.
}
\keyword{internal}
