% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filepath.ends.with}
\alias{filepath.ends.with}
\title{Check whether filepath ends with extension.}
\usage{
filepath.ends.with(filepath, extensions)
}
\arguments{
\item{filepath}{string. Path to a file, including filename and extension.}

\item{extensions}{list of strings. A list of suffixes to check. Case does not matter. Example: \code{extensions=c('.gz', '.mgz')}.}
}
\value{
logical, whether the filepath end with one of the extensions.
}
\description{
Check whether filepath ends with extension.
}
\keyword{internal}
