% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{finite.rows}
\alias{finite.rows}
\title{Check which rows of a matrix consist of finite values only.}
\usage{
finite.rows(mat)
}
\arguments{
\item{mat}{numeric matrix.}
}
\value{
logical vector with one entry per row.
}
\description{
Check which rows of a matrix consist of finite values only.
}
\keyword{internal}
