% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate3d.R
\name{flip2D}
\alias{flip2D}
\title{Flip a 2D matrix.}
\usage{
flip2D(slice, how = "horizontally")
}
\arguments{
\item{slice}{a 2D matrix}

\item{how}{character string, one of 'vertically' / 'v' or 'horizontally' / 'h'. Note that flipping \emph{horizontally} means that the image will be mirrored along the central \emph{vertical} axis. If \code{NULL} is passed, the passed value is returned unaltered.}
}
\value{
2D matrix, the flipped matrix.
}
\description{
Flip a 2D matrix.
}
