% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{fread3}
\alias{fread3}
\title{Read 3-byte integer.}
\usage{
fread3(filehandle)
}
\arguments{
\item{filehandle:}{file handle}
}
\value{
integer: The read integer.
}
\description{
Read a 3-byte integer from a binary file handle. Advances the pointer accordingly.
}
\keyword{internal}
