% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fs.surface.to.tmesh3d}
\alias{fs.surface.to.tmesh3d}
\title{Get an rgl tmesh3d instance from a brain surface mesh.}
\usage{
fs.surface.to.tmesh3d(surface)
}
\arguments{
\item{surface}{an fs.surface instance, as returned \code{freesurferformats::read.fs.surface}.}
}
\value{
a \code{tmesh3d} instance representing the surface, see \code{rgl::tmesh3d} for details. It has classes \code{mesh3d} and \code{shape3d}.
}
\description{
Convert \code{fs.surface} to \code{tmesh} without the \code{rgl} package.
}
