% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{fs.tracts.count}
\alias{fs.tracts.count}
\title{Get the number of tracts.}
\usage{
fs.tracts.count(tracts)
}
\arguments{
\item{tracts}{an \code{fs.tracts} instance.}
}
\value{
integer, the number of tracts.
}
\description{
Get the number of tracts.
}
\keyword{internal}
