% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{fs.tracts.lengths}
\alias{fs.tracts.lengths}
\title{Get the number of points of each tract.}
\usage{
fs.tracts.lengths(tracts)
}
\arguments{
\item{tracts}{an \code{fs.tracts} instance, as returned in the \code{tracks}
entry of \code{read.dti.tck} or \code{read.dti.trk}.}
}
\value{
integer vector with one entry per tract.
}
\description{
Returns one integer per tract, the number of points it consists
of. The coordinates of tract \code{i} are the rows
\code{(cumsum(c(1, lengths))[i]):(cumsum(lengths)[i])} of
\code{fs.tracts.coords()}.
}
\examples{
\dontrun{
tck <- read.dti.tck("brain.tck");
lengths <- fs.tracts.lengths(tck$tracks);
mean(lengths);
}

}
