% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{fs.tracts.point.count}
\alias{fs.tracts.point.count}
\title{Get the total number of points of all tracts.}
\usage{
fs.tracts.point.count(tracts)
}
\arguments{
\item{tracts}{an \code{fs.tracts} instance.}
}
\value{
integer, the total number of points.
}
\description{
Get the total number of points of all tracts.
}
\keyword{internal}
