% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh2nii.R
\name{fs.volume.from.oro.nifti}
\alias{fs.volume.from.oro.nifti}
\title{Turn an `oro.nifti` instance into an `fs.volume` instance.}
\usage{
fs.volume.from.oro.nifti(nifti_img)
}
\arguments{
\item{nifti_img}{instance of class `nifti` from the `oro.nifti` package}
}
\value{
an `fs.volume` instance
}
\description{
This is work in progress. Currently only few datatypes are supported, and the sform needs to be present in the nifti instance.
}
\examples{
\dontrun{
   nii_file = "~/data/tim_only/tim/mri/brain.nii";  # mri_convert brain.mgz brain.nii
   nii_img = oro.nifti::readNIfTI(nii_file);
   fsvolume = fs.volume.from.oro.nifti(nii_img);
}

}
