% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_curv.R
\name{fwrite3}
\alias{fwrite3}
\title{Write 3-byte integer.}
\usage{
fwrite3(filehandle, data)
}
\arguments{
\item{filehandle}{file handle (connection)}

\item{data}{number to write}
}
\description{
Write a 3-byte integer to a binary file handle.
}
\keyword{internal}
