% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_xml_tools.R
\name{giftixml_add_labeltable_from_annot}
\alias{giftixml_add_labeltable_from_annot}
\title{Add a label tabel from an annotation to a GIFTI XML tree.}
\usage{
giftixml_add_labeltable_from_annot(xmltree, annot)
}
\arguments{
\item{xmltree}{an XML tree from xml2, typically the return value from \code{\link[freesurferformats]{gifti_xml}}.}

\item{annot}{an fs.annotation, the included data will be used to compute the LabelTable node}
}
\value{
XML tree from xml2, the modified tree with the LabelTable added below the root node.
}
\description{
Computes the LabelTable XML node for the given annotation and adds it to the XML tree.
}
\keyword{internal}
