% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_xml_tools.R
\name{giftixml_add_labeltable_posneg}
\alias{giftixml_add_labeltable_posneg}
\title{Add a standard label tabel for to a GIFTI XML tree.}
\usage{
giftixml_add_labeltable_posneg(xmltree)
}
\arguments{
\item{xmltree}{an XML tree from xml2, typically the return value from \code{\link[freesurferformats]{gifti_xml}}.}
}
\value{
XML tree from xml2, the modified tree with the LabelTable added below the root node.
}
\description{
This label table is suitable for labels in the FreeSurfer sense, i.e., if only a positive label (1) and a negative label (0) exist in the label data.
}
\keyword{internal}
