% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_transform.R
\name{guess.writable.transform.format}
\alias{guess.writable.transform.format}
\title{Determine the format of a transformation file to write.}
\usage{
guess.writable.transform.format(filepath)
}
\arguments{
\item{filepath}{character string, the full path of the file to write.}
}
\value{
character string, the file format.
}
\description{
Guess the transformation file format from the file name extension, for the formats this package
can write.
}
\keyword{internal}
