% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_pandoc}
\alias{has_pandoc}
\title{Check for pandoc availability on system.}
\usage{
has_pandoc()
}
\value{
logical, whether Pandoc is available.
}
\description{
Check for pandoc availability on system.
}
\keyword{internal}
