% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{int.to.col.brainvoyager}
\alias{int.to.col.brainvoyager}
\title{Convert 32 bit integer to RGB color as described in Brainvoyager SRF file spec.}
\usage{
int.to.col.brainvoyager(int_val)
}
\arguments{
\item{int_val}{the 32 bit integer}
}
\value{
an rgb color
}
\description{
Convert 32 bit integer to RGB color as described in Brainvoyager SRF file spec.
}
\keyword{internal}
