% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvoyager.R
\name{is.bvsmp}
\alias{is.bvsmp}
\title{Check whether object is a bvsmp instance.}
\usage{
is.bvsmp(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
TRUE if its argument is an bvsmp instane (that is, has "bvsmp" amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is a bvsmp instance.
}
