% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_annot.R
\name{is.fs.annot}
\alias{is.fs.annot}
\title{Check whether object is an fs.annot}
\usage{
is.fs.annot(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
TRUE if its argument is a brain surface annotation (that is, has "fs.annot" amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is an fs.annot
}
