% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_label.R
\name{is.fs.label}
\alias{is.fs.label}
\title{Check whether object is an fs.label}
\usage{
is.fs.label(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
TRUE if its argument is a brain surface label (that is, has \code{fs.label} amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is an fs.label
}
