% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{is.fs.tracts}
\alias{is.fs.tracts}
\title{Check whether an object is an fs.tracts instance.}
\usage{
is.fs.tracts(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
logical, TRUE if \code{x} is an \code{fs.tracts} instance.
}
\description{
Check whether an object is an fs.tracts instance.
}
\examples{
is.fs.tracts("not tracts");

}
