% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{is.fs.volume}
\alias{is.fs.volume}
\title{Check whether object is an fs.volume}
\usage{
is.fs.volume(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
TRUE if its argument is a brain volume (that is, has "fs.volume" amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is an fs.volume
}
