% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvis_affine.R
\name{is.identity.matrix}
\alias{is.identity.matrix}
\title{Check whether a matrix is (close to) the identity.}
\usage{
is.identity.matrix(mat, tolerance = 1e-06)
}
\arguments{
\item{mat}{numeric matrix.}

\item{tolerance}{numeric, the tolerance for the comparison.}
}
\value{
logical.
}
\description{
Check whether a matrix is (close to) the identity.
}
\keyword{internal}
