% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{is.mghheader}
\alias{is.mghheader}
\title{Check whether object is an mghheader}
\usage{
is.mghheader(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
TRUE if its argument is an MGH header (that is, has "mghheader" amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is an mghheader
}
