% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{is.quakemodel}
\alias{is.quakemodel}
\title{Check whether object is Quake 1 or 2 alias model}
\usage{
is.quakemodel(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Check whether object is Quake 1 or 2 alias model
}
