% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{is.quakemodel_md2}
\alias{is.quakemodel_md2}
\title{Check whether object is Quake 2 MD2 model}
\usage{
is.quakemodel_md2(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Check whether object is Quake 2 MD2 model
}
