% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{is.quakemodel_mdl}
\alias{is.quakemodel_mdl}
\title{Check whether object is Quake 1 MDL model}
\usage{
is.quakemodel_mdl(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Check whether object is Quake 1 MDL model
}
