% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{length.fs.tracts}
\alias{length.fs.tracts}
\title{Number of tracts in an fs.tracts instance.}
\usage{
\method{length}{fs.tracts}(x)
}
\arguments{
\item{x}{an \code{fs.tracts} instance.}
}
\value{
integer, the number of tracts.
}
\description{
Number of tracts in an fs.tracts instance.
}
