% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{linesplit.fixed}
\alias{linesplit.fixed}
\title{Split a string into fixed-length parts.}
\usage{
linesplit.fixed(
  cline,
  length_per_part,
  num_parts_expected = NULL,
  error_tag = NULL
)
}
\arguments{
\item{cline}{character string, the input line}

\item{length_per_part}{integer, number of characters per part}

\item{num_parts_expected}{integer, the number of parts. Leave at NULL if this is not known.}

\item{error_tag}{optional character string, how to identify the line in a parsing error message. Could be the line number, or whatever. Only relevant if 'num_parts_expected' is not matched.}
}
\description{
Split a string into fixed-length parts.
}
\keyword{internal}
