% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{lookup.q1.normals}
\alias{lookup.q1.normals}
\title{Lookup Quake I normals by index.}
\usage{
lookup.q1.normals(normal_indices)
}
\arguments{
\item{normal_indices}{integer vector of length n, the normal indices (0-based).}
}
\value{
n x 3 matrix of doubles, the normals
}
\description{
Lookup Quake I normals by index.
}
\keyword{internal}
