% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_volume.R
\name{m44_to_quaternion}
\alias{m44_to_quaternion}
\title{Compute quaternion representation of a rotation from a 4x4 rotation matrix.}
\usage{
m44_to_quaternion(m)
}
\arguments{
\item{m}{the input 4x4 matrix encoding the rotation, with homogeneous column \link{0,0,0,1}.}
}
\value{
numeric vector of length 4, the quaternion representation (qw, qx, qy, qz).
}
\description{
Compute quaternion representation of a rotation from a 4x4 rotation matrix.
}
\keyword{internal}
