% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{merge.bbox}
\alias{merge.bbox}
\title{Merge two bounding boxes.}
\usage{
\method{merge}{bbox}(bbox1, bbox2)
}
\arguments{
\item{bbox1}{numeric vector of length 6 or NULL.}

\item{bbox2}{numeric vector of length 6 or NULL.}
}
\value{
numeric vector of length 6, the union of the two boxes.
}
\description{
Merge two bounding boxes.
}
\keyword{internal}
