% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.centervoxelRAS.from.firstvoxelRAS}
\alias{mghheader.centervoxelRAS.from.firstvoxelRAS}
\title{Compute RAS coords of center voxel.}
\usage{
mghheader.centervoxelRAS.from.firstvoxelRAS(header, first_voxel_RAS)
}
\arguments{
\item{header}{Header of the mgh datastructure, as returned by \code{\link[freesurferformats]{read.fs.mgh}}. The \code{c_r}, \code{c_a} and \code{c_s} values in the header do not matter of course, they are what is computed by this function.}

\item{first_voxel_RAS}{numerical vector of length 3, the RAS coordinate of the first voxel in the volume. The first voxel is the voxel with \verb{CRS=1,1,1} in R, or \verb{CRS=0,0,0} in C/FreeSurfer. This value is also known as \emph{P0 RAS}.}
}
\value{
numerical vector of length 3, the RAS coordinate of the center voxel. Also known as \emph{CRAS} or \emph{center RAS}.
}
\description{
Compute RAS coords of center voxel.
}
