% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.crs.orientation}
\alias{mghheader.crs.orientation}
\title{Compute MGH volume orientation string.}
\usage{
mghheader.crs.orientation(header)
}
\arguments{
\item{header}{Header of the mgh datastructure, as returned by \code{\link[freesurferformats]{read.fs.mgh}}.}
}
\value{
character string of length 3, one uppercase letter per axis. Each of the three position is a letter from the alphabet: \verb{LRISAP?}. The meaning is \code{L} for left, \code{R} for right, \code{I} for inferior, \code{S} for superior, \code{P} for posterior, \code{A} for anterior. If the direction cannot be computed, all three characters are \verb{?} for unknown. Of course, each axis (\code{L/R}, \code{I/S}, \code{A/P}) is only represented once in the string.
}
\description{
Compute MGH volume orientation string.
}
