% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.primary.slice.direction}
\alias{mghheader.primary.slice.direction}
\title{Compute MGH primary slice direction}
\usage{
mghheader.primary.slice.direction(header)
}
\arguments{
\item{header}{Header of the mgh datastructure, as returned by \code{\link[freesurferformats]{read.fs.mgh}}.}
}
\value{
character string, the slice direction. One of 'sagittal', 'coronal', 'axial' or 'unknown'.
}
\description{
Compute MGH primary slice direction
}
