% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.ras2vox.xform}
\alias{mghheader.ras2vox.xform}
\title{Compute ras2vox xform}
\usage{
mghheader.ras2vox.xform(header)
}
\arguments{
\item{header}{the MGH header}
}
\value{
4x4 numerical matrix, the tranformation matrix
}
\description{
Compute ras2vox xform
}
