% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader}
\alias{mghheader}
\title{Constructor to init MGH header instance.}
\usage{
mghheader(dims, mri_dtype_code)
}
\arguments{
\item{dims}{integer vector of length 4, the header dimensions. Example: \code{c(256L, 256L, 256L, 1L)}.}

\item{mri_dtype_code}{integer, a valid MRI datatype. See \code{\link[freesurferformats]{translate.mri.dtype}}.}
}
\value{
a named list representing the header
}
\description{
Constructor to init MGH header instance.
}
\keyword{internal}
