% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.update.from.vox2ras}
\alias{mghheader.update.from.vox2ras}
\title{Update mghheader fields from vox2ras matrix.}
\usage{
mghheader.update.from.vox2ras(header, vox2ras)
}
\arguments{
\item{header}{Header of the mgh datastructure, as returned by \code{\link[freesurferformats]{read.fs.mgh}}.}

\item{vox2ras}{4x4 numerical matrix, the vox2ras transformation matrix.}
}
\value{
a named list representing the header
}
\description{
Update mghheader fields from vox2ras matrix.
}
