% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.vox2ras.xform}
\alias{mghheader.vox2ras.xform}
\title{Compute vox2ras xform}
\usage{
mghheader.vox2ras.xform(header)
}
\arguments{
\item{header}{the MGH header}
}
\value{
4x4 numerical matrix, the tranformation matrix
}
\description{
Compute vox2ras xform
}
