% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{mghheader.vox2rastkreg}
\alias{mghheader.vox2rastkreg}
\title{Compute vox2ras tkreg matrix from basic MGH header fields.}
\usage{
mghheader.vox2rastkreg(header)
}
\arguments{
\item{header}{the MGH header}
}
\description{
This is also known as the 'tkreg' vox2ras.
}
