% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.vox2vox}
\alias{mghheader.vox2vox}
\title{Compute vox2vox matrix between two volumes.}
\usage{
mghheader.vox2vox(header_from, header_to)
}
\arguments{
\item{header_from}{the MGH header of the source volume}

\item{header_to}{the MGH header of the target volume}
}
\value{
4x4 numerical matrix, the tranformation matrix
}
\description{
Compute vox2vox matrix between two volumes.
}
