% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_mgh.R
\name{mri_dtype_numbytes}
\alias{mri_dtype_numbytes}
\title{Get size of MRI dtype in bytes.}
\usage{
mri_dtype_numbytes(mri_dtype_code)
}
\arguments{
\item{mri_dtype_code}{integer, the MRI data type code. See \code{\link[freesurferformats]{translate.mri.dtype}}.}
}
\value{
integer, the number of bytes
}
\description{
Get size of MRI dtype in bytes.
}
\keyword{internal}
