% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti2.R
\name{ni2header.for.data}
\alias{ni2header.for.data}
\title{Create NIFTI v2 header suitable for given data.}
\usage{
ni2header.for.data(niidata)
}
\arguments{
\item{niidata}{array of numeric (integer or double) data, can have up to 7 dimensions.}
}
\value{
a NIFTI v2 header (see \code{\link{ni2header.template}}) in which the datatype, bitpix, dim and dim_raw fields have been set to values suitable for the given data. Feel free to change the other fields.
}
\description{
Create NIFTI v2 header suitable for given data.
}
