% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.dtypebitpix.info.from.mgh.dtype}
\alias{nifti.dtypebitpix.info.from.mgh.dtype}
\title{Compute the 'datatype' and 'bitpix' fields used in the NIFTI1 header from an MGH/MGZ datatype code.}
\usage{
nifti.dtypebitpix.info.from.mgh.dtype(mgh_dtype_code)
}
\arguments{
\item{mgh_dtype_code}{integer, the MGH/MGZ datatype code (as returned by \code{translate.mri.dtype}).}
}
\value{
named list with entries: \code{datatype} and \code{bitpix} containing the translated data for the respective NIfTI-1 header fields.
}
\description{
Compute the 'datatype' and 'bitpix' fields used in the NIFTI1 header from an MGH/MGZ datatype code.
}
\note{
This is useful to compute a NIFTI v1 header from an MGH header.
}
\keyword{internal}
