% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_common.R
\name{nifti.field.check.length}
\alias{nifti.field.check.length}
\title{Check whether vector has expected length.}
\usage{
nifti.field.check.length(niiheader, fields, dlength)
}
\arguments{
\item{niiheader}{named list, representing a NIFTI v1 or v2 header}

\item{fields}{vector of character string, the field names to check}

\item{dlength}{integer, the expected length of all fields}
}
\value{
logical, whether the checks were okay
}
\description{
Check whether vector has expected length.
}
\keyword{internal}
