% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_common.R
\name{nifti.field.check.nchar.max}
\alias{nifti.field.check.nchar.max}
\title{Check whether character string fields have less than or equal to expected length.}
\usage{
nifti.field.check.nchar.max(niiheader, fields, dlength)
}
\arguments{
\item{niiheader}{named list, representing a NIFTI v1 or v2 header}

\item{fields}{vector of character string, the field names to check}

\item{dlength}{integer, the max length of all fields}
}
\value{
logical, whether the checks were okay
}
\description{
Check whether character string fields have less than or equal to expected length.
}
\keyword{internal}
