% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti1.R
\name{nifti.file.version}
\alias{nifti.file.version}
\title{Determine NIFTI file version information and whether file is a NIFTI file.}
\usage{
nifti.file.version(filepath)
}
\arguments{
\item{filepath}{path to a file in NIFTI v1 or v2 format.}
}
\value{
integer, the NIFTI file version. One if \code{1} for NIFTI v1 files, \code{2} for NIFTI v2 files, or \code{NULL} if the file is not a NIFTI file.
}
\description{
Determine NIFTI file version information and whether file is a NIFTI file.
}
