% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_common.R
\name{nifti.header.check}
\alias{nifti.header.check}
\title{Perform basic sanity checks on NIFTI header data. These are in no way meant to be exhaustive.}
\usage{
nifti.header.check(niiheader, nifti_version = 1L)
}
\arguments{
\item{niiheader}{named list, the NIFTI header.}

\item{nifti_version}{integer, one of 1L or 2L. The NIFTI format version.}
}
\value{
logical, whether the check was okay
}
\description{
Perform basic sanity checks on NIFTI header data. These are in no way meant to be exhaustive.
}
