% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_nifti_hack.R
\name{nifti.header.fshack}
\alias{nifti.header.fshack}
\title{Read NIFTI v1 header from file with FreeSurfer hack.}
\usage{
nifti.header.fshack(filepath)
}
\arguments{
\item{filepath}{path to a NIFTI v1 file (single file version), which can contain the FreeSurfer hack.}
}
\value{
named list with NIFTI 1 header fields.
}
\description{
Read NIFTI v1 header from file with FreeSurfer hack.
}
\note{
The FreeSurfer hack is a non-standard way to save long vectors (one dimension greater than 32k entries) in NIFTI v1 files. Files with this hack are produced when converting MGH or MGZ files containing such long vectors with the FreeSurfer 'mri_convert' tool.
}
