% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.info.from.oro.instance}
\alias{nifti.info.from.oro.instance}
\title{Extract the header information and the data of an \code{oro.nifti} instance into a plain list.}
\usage{
nifti.info.from.oro.instance(nifti_img)
}
\arguments{
\item{nifti_img}{an instance of class \code{nifti} from the \code{oro.nifti} package.}
}
\value{
named list, see \code{\link{nifti.info.from.file}}.
}
\description{
Extract the header information and the data of an \code{oro.nifti} instance into a plain list.
}
\note{
The data scaling fields of the instance are not applied here, they are expected to be applied already:
instances read by \code{oro.nifti::readNIfTI} are rescaled unless \code{rescale_data = FALSE} was used.
}
\seealso{
\code{\link{nifti.info.from.file}}
}
\keyword{internal}
