% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.resolve.filepath}
\alias{nifti.resolve.filepath}
\title{Resolve the path of a NIFTI file which is given without a file extension.}
\usage{
nifti.resolve.filepath(filepath)
}
\arguments{
\item{filepath}{character string, the path to a NIFTI file.}
}
\value{
character string, the path of an existing file. If \code{filepath} itself does not exist, the usual
NIFTI file extensions are appended to it and the first existing file is returned.
}
\description{
Resolve the path of a NIFTI file which is given without a file extension.
}
\keyword{internal}
