% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.transform.type.name}
\alias{nifti.transform.type.name}
\title{Get the name of the transform type from a form code.}
\usage{
nifti.transform.type.name(form_code)
}
\arguments{
\item{form_code}{integer, the value retrieved from the \code{sform_code} or the \code{qform_code} NIFTI header fields}
}
\value{
character string, the meaning of the code. Usually this expresses to what the data will be aligned after application of the vox2ras transformation method. (The type of transformation to perform in order to achieve this alignment depends on whether the value was retrieved from the \code{sform} or the \code{qform} field and does not matter here.)
}
\description{
The form code is a code stored in the \code{sform_code} and/or \code{qform_code} NIFTI header fields.
}
\keyword{internal}
