% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti2_extensions.R
\name{nifti2.extension.content}
\alias{nifti2.extension.content}
\title{Get the payload of a NIFTI v2 header extension as raw bytes.}
\usage{
nifti2.extension.content(extension, strip_nul = TRUE)
}
\arguments{
\item{extension}{a NIFTI v2 header extension, a named list with entries 'ecode' and 'content', see \code{\link{nifti2.extension}}.}

\item{strip_nul}{logical, whether to remove NUL bytes from the payload. This is required to convert the payload to a character string.}
}
\value{
a raw vector, the payload of the extension.
}
\description{
Get the payload of a NIFTI v2 header extension as raw bytes.
}
\seealso{
Other nifti2 extensions:
\code{\link[=nifti2.extension]{nifti2.extension()}},
\code{\link[=nifti2.extension.size]{nifti2.extension.size()}},
\code{\link[=nifti2.extension.text]{nifti2.extension.text()}},
\code{\link[=nifti2.get.extension]{nifti2.get.extension()}}
}
\concept{nifti2 extensions}
\keyword{internal}
