% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_niftiv2.R
\name{nifti2.header}
\alias{nifti2.header}
\title{Read NIFTI v2 header from file.}
\usage{
nifti2.header(filepath)
}
\arguments{
\item{filepath}{path to a NIFTI v2 file.}
}
\value{
named list with NIFTI 2 header fields.
}
\description{
Read NIFTI v2 header from file.
}
