% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti2_extensions.R
\name{nifti2.read.extensions}
\alias{nifti2.read.extensions}
\title{Read the header extensions of a NIFTI v2 file.}
\usage{
nifti2.read.extensions(fh, available, endian = "little")
}
\arguments{
\item{fh}{filehandle, a connection to a NIFTI v2 file, positioned at the first byte after the fixed-size header (i.e., at the extension flag bytes).}

\item{available}{integer, the number of bytes available for header extensions, i.e., \code{vox_offset} minus the 544 bytes of the fixed-size header and the extension flag.}

\item{endian}{character string, the endianness of the file, either 'little' or 'big'.}
}
\value{
list of header extensions, see \code{\link{nifti2.extension}}. Empty list if the file has no extensions.
}
\description{
Read the header extensions of a NIFTI v2 file.
}
\keyword{internal}
