% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti2_extensions.R
\name{nifti2.strip.nul}
\alias{nifti2.strip.nul}
\title{Remove NUL bytes from a raw vector.}
\usage{
nifti2.strip.nul(rawdata)
}
\arguments{
\item{rawdata}{raw vector or integer vector with values in range 0 to 255.}
}
\value{
raw vector, the input without any NUL bytes.
}
\description{
Remove all NUL bytes from a raw vector. This is needed for the payload of NIFTI v2 header extensions, which are padded with NUL bytes, and for the payload of MATLAB files, but the function is not specific to any format.
}
\keyword{internal}
