% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti2_extensions.R
\name{nifti2.write.extensions}
\alias{nifti2.write.extensions}
\title{Write the header extensions of a NIFTI v2 file.}
\usage{
nifti2.write.extensions(fh, extensions, endian = "little")
}
\arguments{
\item{fh}{filehandle, a connection to a NIFTI v2 file, positioned at the first byte after the fixed-size header (i.e., where the extension flag bytes go).}

\item{extensions}{list of header extensions, see \code{\link{nifti2.extension}}. An empty list writes the extension flag for 'no extensions present'.}

\item{endian}{character string, the endianness of the file, either 'little' or 'big'.}
}
\value{
integer, the number of bytes written (the extension flag bytes plus all extensions). The data must start at \verb{544 + <return value>}.
}
\description{
Write the header extensions of a NIFTI v2 file.
}
\keyword{internal}
