% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrrd.R
\name{nrrd.parse.matrix}
\alias{nrrd.parse.matrix}
\title{Parse a NRRD matrix value like '(1,0,0) (0,1,0) (0,0,1)'.}
\usage{
nrrd.parse.matrix(value)
}
\arguments{
\item{value}{character string, the field value.}
}
\value{
numeric matrix, or NULL for 'none'.
}
\description{
Parse a NRRD matrix value like '(1,0,0) (0,1,0) (0,0,1)'.
}
\keyword{internal}
