% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrrd.R
\name{nrrd.parse.quoted.list}
\alias{nrrd.parse.quoted.list}
\title{Parse a quoted NRRD string list.}
\usage{
nrrd.parse.quoted.list(value)
}
\arguments{
\item{value}{character string, the field value, e.g. 'left right' or '"some label"'.}
}
\value{
character vector.
}
\description{
Parse a quoted NRRD string list.
}
\keyword{internal}
