% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrrd.R
\name{nrrd.parse.vector}
\alias{nrrd.parse.vector}
\title{Parse a NRRD vector value like '(1,2,3)'.}
\usage{
nrrd.parse.vector(value)
}
\arguments{
\item{value}{character string, the field value.}
}
\value{
numeric vector, or NULL for 'none'.
}
\description{
Parse a NRRD vector value like '(1,2,3)'.
}
\keyword{internal}
