% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{open.maybe.gzip}
\alias{open.maybe.gzip}
\title{Open a connection that transparently handles gzip compression.}
\usage{
\method{open}{maybe.gzip}(filepath, gzipped, mode = "rb")
}
\arguments{
\item{filepath}{character string, path to the file.}

\item{gzipped}{logical, whether the file is gzip-compressed (see
\code{is.gzip.file}).}

\item{mode}{character string, 'r' for text or 'rb' for binary.}
}
\value{
a connection, call \code{close()} on it when done.
}
\description{
Open a connection that transparently handles gzip compression.
}
\keyword{internal}
