% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvis_affine.R
\name{ornt.transform}
\alias{ornt.transform}
\title{Determine the orientation transform between two orientations.}
\usage{
ornt.transform(start_ornt, end_ornt)
}
\arguments{
\item{start_ornt}{numeric matrix with 2 columns, the initial orientation.}

\item{end_ornt}{numeric matrix with 2 columns, the target orientation.}
}
\value{
numeric matrix with 2 columns, the orientation that maps
\code{start_ornt} onto \code{end_ornt}.
}
\description{
Determine the orientation transform between two orientations.
}
\keyword{internal}
