% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvis_affine.R
\name{ornt2axcodes}
\alias{ornt2axcodes}
\title{Convert an orientation array to axis codes.}
\usage{
ornt2axcodes(ornt)
}
\arguments{
\item{ornt}{numeric matrix with 2 columns, see \code{\link{io.orientation}}.}
}
\value{
character vector with one code per input axis, drawn from
\code{c('L','R','P','A','I','S')}. Dropped axes become NA.
}
\description{
Convert an orientation array to axis codes.
}
\keyword{internal}
