% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_surface.R
\name{ply.header.lines}
\alias{ply.header.lines}
\title{Generate PLY format header lines}
\usage{
ply.header.lines(num_verts, num_faces, use_vertex_colors)
}
\description{
Generate PLY format header lines
}
\keyword{internal}
