% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{predefined.mdl.normals}
\alias{predefined.mdl.normals}
\title{Return list of pre-defined Quake I normals.}
\usage{
predefined.mdl.normals()
}
\value{
n x 3 matrix of doubles, the normals. Hardcoded.
}
\description{
Return list of pre-defined Quake I normals.
}
\keyword{internal}
