% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{print.dti.track.iterator}
\alias{print.dti.track.iterator}
\title{Print a tract file iterator.}
\usage{
\method{print}{dti.track.iterator}(x, ...)
}
\arguments{
\item{x}{a \code{dti.track.iterator} instance.}

\item{...}{ignored.}
}
\value{
the iterator, invisibly.
}
\description{
Print a tract file iterator.
}
