% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_patch.R
\name{print.fs.patch}
\alias{print.fs.patch}
\title{Print description of a brain surface patch.}
\usage{
\method{print}{fs.patch}(x, ...)
}
\arguments{
\item{x}{brain surface patch with class \code{fs.patch}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print description of a brain surface patch.
}
