% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{print.fs.surface}
\alias{print.fs.surface}
\title{Print description of a brain surface.}
\usage{
\method{print}{fs.surface}(x, ...)
}
\arguments{
\item{x}{brain surface with class \code{fs.surface}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print description of a brain surface.
}
