% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{print.fs.tracts}
\alias{print.fs.tracts}
\title{Print an fs.tracts instance.}
\usage{
\method{print}{fs.tracts}(x, ...)
}
\arguments{
\item{x}{an \code{fs.tracts} instance.}

\item{...}{ignored.}
}
\value{
the instance, invisibly.
}
\description{
Print an fs.tracts instance.
}
