% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{print.fs.volume}
\alias{print.fs.volume}
\title{Print description of a brain volume.}
\usage{
\method{print}{fs.volume}(x, ...)
}
\arguments{
\item{x}{brain volume with class \code{fs.volume}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print description of a brain volume.
}
