% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_analyze.R
\name{read.analyze.data}
\alias{read.analyze.data}
\title{Read ANALYZE 7.5 data from file.}
\usage{
read.analyze.data(filepath, header = NULL, drop_empty_dims = TRUE)
}
\arguments{
\item{filepath}{character string, the path to the \code{.hdr} file. The base name without the extension is accepted as
well, see \code{\link{analyze.pair.files}}.}

\item{header}{optional ANALYZE 7.5 header as returned by \code{\link{read.analyze.header}}. It will be read from
the file automatically if left at \code{NULL}.}

\item{drop_empty_dims}{logical, whether to drop empty dimensions in the returned data array, see
\code{\link{drop}}.}
}
\value{
the data in the \code{.img} file of the pair, as an array (or a vector for 1D data). Note that the fields of
the header are not applied in any way: the data are returned exactly as they are stored in the file, in the
storage order of the format (the first dimension varies fastest, which is also the order used by R and by the
MGH/MGZ format), and the \code{funused1} field that SPM uses as a scale factor is not applied (see
\code{\link{read.fs.volume.analyze}} if you want that).
}
\description{
Read ANALYZE 7.5 data from file.
}
\examples{
hdrfile <- system.file("extdata", "analyze", "tiny_u8.hdr",
  package = "freesurferformats", mustWork = TRUE
)
data <- read.analyze.data(hdrfile)
dim(data)

}
\seealso{
\code{\link{read.analyze.header}}
}
