% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_track.R
\name{read.dti.track}
\alias{read.dti.track}
\title{Read DTI tracking data from file in MRtrix 'TCK'/'TSF' format.}
\usage{
read.dti.track(filepath)
}
\arguments{
\item{filepath}{character string, path to the \code{TCK} or \code{TSF} file to read.}
}
\value{
named list with entries 'header' and 'tracks'. For \code{TCK}, the tracks are organized into a list of matrices.
Each n x 3 matrix represents the coordinates for the n points of one track, the values in each row are the xyz-coords.
For \code{TSF}, the scalar data are available in 2 representations:
'merged': a vector of all values (requires external knowledge on track borders), and
'scalar_list': organized into a list of vectors. Each vector represents the values for the points of one track.
}
\description{
Read DTI tracking data from file in MRtrix 'TCK'/'TSF' format.
}
\note{
\code{read.dti.track()} supersedes \code{read.dti.tck()} and \code{read.dti.tsf()}.

\url{https://mrtrix.readthedocs.io/en/latest/getting_started/image_data.html#tracks-file-format-tck}
\url{https://mrtrix.readthedocs.io/en/latest/getting_started/image_data.html#track-scalar-file-format-tsf}
}
\examples{
\dontrun{
 tckf = "~/simple.tck"
 tck = read.dti.track(tckf)
 tsff = "~/simple.tsf"
 tsf = read.dti.track(tsff)
}

}
\seealso{
\code{read.dti.tck}
}
