% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_trk.R
\name{read.dti.trk.header}
\alias{read.dti.trk.header}
\title{Read the header of a TrackVis TRK file.}
\usage{
read.dti.trk.header(filepath, shift_origin = TRUE)
}
\arguments{
\item{filepath}{character string, path to a file in TRK format. Gzip-compressed
files (typically named \code{.trk.gz}) are supported, the compression is
detected from the file content, not from the file name.}

\item{shift_origin}{logical, whether to compute the corrected \code{vox2ras}
matrix, see \code{\link{read.dti.trk}}.}
}
\value{
named list, the header. It contains all fields of the TRK header as
described at \code{http://trackvis.org/docs/?subsect=fileformat}, plus
\code{vox2ras} (the matrix stored in the file) and, if
\code{shift_origin} is \code{TRUE}, \code{vox2ras_corrected}.
}
\description{
Reads and parses the 1000 byte header of a TRK file without
touching the track data. This is cheap even for huge tractograms and can be
used to inspect a file before deciding whether to read its tracks.
}
\examples{
\dontrun{
hdr <- read.dti.trk.header("brain.trk");
hdr$n_count;      # number of tracks stated in the header
hdr$voxel_order;
}

}
