% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{read.dti.tsf.header}
\alias{read.dti.tsf.header}
\title{Read the header of a file in MRtrix 'TSF' format.}
\usage{
read.dti.tsf.header(filepath)
}
\arguments{
\item{filepath}{character string, path to the \code{TCK} or \code{TSF} file
to read. Gzip-compressed files (i.e., \code{.tck.gz}) are supported and the
compression is detected from the file content, not the file name.}
}
\value{
named list, the parsed header.
}
\description{
Reads only the ASCII header of the file, see
\code{\link{read.dti.tck.header}}.
}
\examples{
\dontrun{
hdr <- read.dti.tsf.header("brain.tsf");
}

}
