% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.element.counts.ply.header}
\alias{read.element.counts.ply.header}
\title{Determine element counts from PLY file header.}
\usage{
read.element.counts.ply.header(ply_lines)
}
\arguments{
\item{ply_lines}{vector character strings, all lines of the PLY file}
}
\description{
Determine element counts from PLY file header.
}
\keyword{internal}
