% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_common.R
\name{read.fixed.char.binary}
\alias{read.fixed.char.binary}
\title{Read fixed length char, possibly containing embedded zeroes, from binary file.}
\usage{
read.fixed.char.binary(filehandle, n, to = "UTF-8")
}
\arguments{
\item{filehandle}{connection to read.}

\item{n}{the number of bytes to read.}

\item{to}{the target character encoding.}
}
\value{
the string in the target encoding.
}
\description{
Read fixed length char, possibly containing embedded zeroes, from binary file.
}
\author{
The original version was written by Brandon Whitcher and Volker Schmid. See the source for the full legal info. The coding style was adapted to freesurferformats and the docs were added by Tim Schäfer.
}
\keyword{internal}
