% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.fs.curv.asc}
\alias{read.fs.curv.asc}
\title{Read morphometry data from ASCII curv format file}
\usage{
read.fs.curv.asc(filepath)
}
\arguments{
\item{filepath}{path to a file in FreeSurfer ASCII curv format. Such a file contains, on each line, the following fields, separated by spaces: vertex_index, vertex_coord_x,  vertex_coord_y,  vertex_coord_z,  morph_data_value.}
}
\value{
numeric vector, the curv data
}
\description{
Read morphometry data from ASCII curv format file
}
\keyword{internal}
