% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.fs.morph.ni2}
\alias{read.fs.morph.ni2}
\title{Read morphometry data from FreeSurfer NIFTI v2 format files.}
\usage{
read.fs.morph.ni2(filepath)
}
\arguments{
\item{filepath}{path to a file in FreeSurfer NIFTI v2 format, potentially with the FreeSurfer hack. See \code{\link{read.nifti2.data}} for details.}
}
\value{
numeric vector, the morphometry data
}
\description{
Read morphometry data from FreeSurfer NIFTI v2 format files.
}
