% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.geo}
\alias{read.fs.surface.geo}
\title{Read GEO format mesh as surface.}
\usage{
read.fs.surface.geo(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input surface file in GEO mesh format.}
}
\value{
named list. The list has the following named entries: "vertices": nx3 double matrix, where n is the number of vertices. Each row contains the x,y,z coordinates of a single vertex. "faces": nx3 integer matrix. Each row contains the vertex indices of the 3 vertices defining the face. WARNING: The indices are returned starting with index 1 (as used in GNU R). Keep in mind that you need to adjust the index (by substracting 1) to compare with data from other software.
}
\description{
This reads meshes from text files in GEO mesh format. This is an ASCII format.
}
\note{
This is a fixed width format.
}
\seealso{
Other mesh functions:
\code{\link[=faces.quad.to.tris]{faces.quad.to.tris()}},
\code{\link[=read.fs.surface]{read.fs.surface()}},
\code{\link[=read.fs.surface.asc]{read.fs.surface.asc()}},
\code{\link[=read.fs.surface.bvsrf]{read.fs.surface.bvsrf()}},
\code{\link[=read.fs.surface.gii]{read.fs.surface.gii()}},
\code{\link[=read.fs.surface.ico]{read.fs.surface.ico()}},
\code{\link[=read.fs.surface.obj]{read.fs.surface.obj()}},
\code{\link[=read.fs.surface.off]{read.fs.surface.off()}},
\code{\link[=read.fs.surface.ply]{read.fs.surface.ply()}},
\code{\link[=read.fs.surface.vtk]{read.fs.surface.vtk()}},
\code{\link[=read.mesh.brainvoyager]{read.mesh.brainvoyager()}},
\code{\link[=read_nisurface]{read_nisurface()}},
\code{\link[=read_nisurfacefile]{read_nisurfacefile()}},
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.asc]{write.fs.surface.asc()}},
\code{\link[=write.fs.surface.byu]{write.fs.surface.byu()}},
\code{\link[=write.fs.surface.gii]{write.fs.surface.gii()}},
\code{\link[=write.fs.surface.mz3]{write.fs.surface.mz3()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh functions}
