% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.obj}
\alias{read.fs.surface.obj}
\title{Read OBJ format mesh as surface.}
\usage{
read.fs.surface.obj(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input surface file in Wavefront object mesh format. Files with non-standard vertex colors (3 additional float fields after the vertex coordinates in order R, G, B) are supported, and the colors will be returned in the field 'vertex_colors' if present.}
}
\value{
named list. The list has the following named entries: "vertices": nx3 double matrix, where n is the number of vertices. Each row contains the x,y,z coordinates of a single vertex. "faces": nx3 integer matrix. Each row contains the vertex indices of the 3 vertices defining the face. WARNING: The indices are returned starting with index 1 (as used in GNU R). Keep in mind that you need to adjust the index (by substracting 1) to compare with data from other software.
}
\description{
This reads meshes from text files in Wavefront OBJ mesh format. This is an ASCII format.
}
\note{
This is a simple but very common mesh format supported by many applications, well suited for export.
}
\seealso{
Other mesh functions:
\code{\link[=faces.quad.to.tris]{faces.quad.to.tris()}},
\code{\link[=read.fs.surface]{read.fs.surface()}},
\code{\link[=read.fs.surface.asc]{read.fs.surface.asc()}},
\code{\link[=read.fs.surface.bvsrf]{read.fs.surface.bvsrf()}},
\code{\link[=read.fs.surface.geo]{read.fs.surface.geo()}},
\code{\link[=read.fs.surface.gii]{read.fs.surface.gii()}},
\code{\link[=read.fs.surface.ico]{read.fs.surface.ico()}},
\code{\link[=read.fs.surface.off]{read.fs.surface.off()}},
\code{\link[=read.fs.surface.ply]{read.fs.surface.ply()}},
\code{\link[=read.fs.surface.vtk]{read.fs.surface.vtk()}},
\code{\link[=read.mesh.brainvoyager]{read.mesh.brainvoyager()}},
\code{\link[=read_nisurface]{read_nisurface()}},
\code{\link[=read_nisurfacefile]{read_nisurfacefile()}},
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.asc]{write.fs.surface.asc()}},
\code{\link[=write.fs.surface.byu]{write.fs.surface.byu()}},
\code{\link[=write.fs.surface.gii]{write.fs.surface.gii()}},
\code{\link[=write.fs.surface.mz3]{write.fs.surface.mz3()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh functions}
