% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_transform.R
\name{read.fs.transform.dat}
\alias{read.fs.transform.dat}
\title{Load transformation matrix from a tkregister dat file.}
\usage{
read.fs.transform.dat(filepath)
}
\arguments{
\item{filepath}{character string, the full path to the transform file.}
}
\value{
an \code{fs.transform} instance. A tkregister matrix maps the movable volume (the source) to the target
volume, so \code{space_in} is 'voxel' and \code{space_out} is 'ras'. It produces RAS coordinates in the tkregister
frame of the target volume, which is why \code{dst} states \code{frame = 'tkreg'}, see
\code{\link{mghheader.vox2ras.tkreg}}. The other entries of the file are kept as the \code{subject},
\code{in_plane_resolution}, \code{between_plane_resolution} and \code{intensity} fields.
}
\description{
Load transformation matrix from a tkregister dat file.
}
\examples{
tf_file <- system.file("extdata", "register.dat",
  package = "freesurferformats",
  mustWork = TRUE
)
transform <- read.fs.transform.dat(tf_file)
transform$matrix

}
\seealso{
Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.itk]{read.fs.transform.itk()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=sm1to0]{sm1to0()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform.to.lps]{transform.to.lps()}},
\code{\link[=transform.to.ras]{transform.to.ras()}},
\code{\link[=transform.to.voxel]{transform.to.voxel()}},
\code{\link[=transform.to.world]{transform.to.world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.dat]{write.fs.transform.dat()}},
\code{\link[=write.fs.transform.fslmat]{write.fs.transform.fslmat()}},
\code{\link[=write.fs.transform.itk]{write.fs.transform.itk()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}},
\code{\link[=write.fs.transform.xfm]{write.fs.transform.xfm()}}
}
\concept{header coordinate space}
