% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.mesh.brainvoyager}
\alias{read.mesh.brainvoyager}
\title{Read Brainvoyager srf format (.srf) mesh.}
\usage{
read.mesh.brainvoyager(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input surface file in SRF mesh format.}
}
\value{
named list of the elements in the file.
}
\description{
Read a mesh and associated data like color and normals from a binary file in BrainVoyager SRF mesh format.
}
\references{
The srf format spec is at https://support.brainvoyager.com/brainvoyager/automation-development/84-file-formats/344-users-guide-2-3-the-format-of-srf-files.
}
\seealso{
Other mesh functions:
\code{\link[=faces.quad.to.tris]{faces.quad.to.tris()}},
\code{\link[=read.fs.surface]{read.fs.surface()}},
\code{\link[=read.fs.surface.asc]{read.fs.surface.asc()}},
\code{\link[=read.fs.surface.bvsrf]{read.fs.surface.bvsrf()}},
\code{\link[=read.fs.surface.geo]{read.fs.surface.geo()}},
\code{\link[=read.fs.surface.gii]{read.fs.surface.gii()}},
\code{\link[=read.fs.surface.ico]{read.fs.surface.ico()}},
\code{\link[=read.fs.surface.obj]{read.fs.surface.obj()}},
\code{\link[=read.fs.surface.off]{read.fs.surface.off()}},
\code{\link[=read.fs.surface.ply]{read.fs.surface.ply()}},
\code{\link[=read.fs.surface.vtk]{read.fs.surface.vtk()}},
\code{\link[=read_nisurface]{read_nisurface()}},
\code{\link[=read_nisurfacefile]{read_nisurfacefile()}},
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.asc]{write.fs.surface.asc()}},
\code{\link[=write.fs.surface.byu]{write.fs.surface.byu()}},
\code{\link[=write.fs.surface.gii]{write.fs.surface.gii()}},
\code{\link[=write.fs.surface.mz3]{write.fs.surface.mz3()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh functions}
