% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti2.R
\name{read.nifti2.header.internal}
\alias{read.nifti2.header.internal}
\title{Read NIFTI v2 header from file.}
\usage{
read.nifti2.header.internal(filepath, little_endian = TRUE)
}
\arguments{
\item{filepath}{path to a NIFTI v2 file.}

\item{little_endian}{internal logical, leave this alone. Endianness will be figured out automatically, messing with this parameter only hurts.}
}
\value{
named list with NIFTI 2 header fields, including the header extensions (field \code{extensions}, see \code{\link{read.nifti2.header}}).
}
\description{
Read NIFTI v2 header from file.
}
\note{
See https://nifti.nimh.nih.gov/pub/dist/data/nifti2/ for test data. Thanks to Anderson Winkler for his post at https://brainder.org/2015/04/03/the-nifti-2-file-format/.
}
\keyword{internal}
