% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{read.quake.md2}
\alias{read.quake.md2}
\title{Read Quake II model in MD2 format.}
\usage{
read.quake.md2(filepath, anim = FALSE)
}
\arguments{
\item{filepath}{character string, the path to the MD2 file}

\item{anim}{logical, whether to load the whole animation (if present). Returns a list of models, the animation frames. If FALSE, only the first frame is returned.}
}
\description{
Read Quake II model in MD2 format.
}
\note{
Ignore this function, it will be moved to a different package.
}
