% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{read.quake.mdl}
\alias{read.quake.mdl}
\title{Read Quake model in MDL format.}
\usage{
read.quake.mdl(filepath, anim = FALSE)
}
\arguments{
\item{filepath}{character string, the path to the MDL file}

\item{anim}{logical, whether to load the whole animation (if present). Returns a list of models, the animation frames. If FALSE, only the first frame is returned.}
}
\description{
Read Quake model in MDL format.
}
\note{
Ignore this function, it will be moved to a different package.
}
\examples{
\dontrun{
   mdlf = "~/data/q1_pak/progs/quaddama.mdl"
   mdl = read.quake.mdl(mdlf);
}

}
