% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nisurface.R
\name{read_nisurfacefile.gifti}
\alias{read_nisurfacefile.gifti}
\title{Read a gifti file as a surface.}
\usage{
\method{read_nisurfacefile}{gifti}(filepath, ...)
}
\arguments{
\item{filepath}{character string, the full path to the input surface file.}

\item{...}{ignored}
}
\value{
an instance of \code{fs.surface}, read from the file. See \code{\link[freesurferformats]{read.fs.surface}} for details. If none of the reader methods succeed, an error is raised.
}
\description{
Read a gifti file as a surface.
}
