% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{readable.files}
\alias{readable.files}
\title{Find files with the given base name and extensions that exist.}
\usage{
readable.files(
  filepath,
  precedence = c(".mgh", ".mgz"),
  error_if_none = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{filepath}{character string, path to a file without extension}

\item{precedence}{vector of character strings, the file extensions to check. Must include the dot (if you expect one).}

\item{error_if_none}{logical, whether to raise an error if none of the files exist}

\item{return_all}{logical, whether to return all readable files instead of just the first one}
}
\value{
character string, the path to the first existing file (or \code{NULL} if none of them exists).
}
\description{
Note that in the current implementation, the case of the filepath and the extension must match.
}
