% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_annot.R
\name{readcolortable}
\alias{readcolortable}
\title{Read binary colortable in v2 format.}
\usage{
readcolortable(fh, ctable_num_entries)
}
\arguments{
\item{filehandle:}{file handle}
}
\value{
list: the color table. Entries are: "num_entries": int, number of brain structures. "struct_names": vector of strings, the names. "table": numeric matrix.
}
\description{
Read a v2 format colortable from a connection to a binary file.
}
\keyword{internal}
