% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate3d.R
\name{rotate3D}
\alias{rotate3D}
\title{Rotate a 3D array in 90 degree steps.}
\usage{
rotate3D(volume, axis = 1L, degrees = 90L)
}
\arguments{
\item{volume}{a 3D image volume}

\item{axis}{positive integer in range 1L..3L or an axis name, the axis to use.}

\item{degrees}{integer, must be a (positive or negative) multiple of 90L.}
}
\value{
a 3D image volume, rotated around the axis. The dimensions may or may not be different from the input image, depending on the rotation angle.
}
\description{
Rotate a 3D array in 90 degree steps along an axis. This leads to an array with different dimensions.
}
\seealso{
Other volume math:
\code{\link[=flip3D]{flip3D()}}
}
\concept{volume math}
