% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate3d.R
\name{rotate90}
\alias{rotate90}
\title{Rotate 2D matrix clockwise in 90 degree steps.}
\usage{
rotate90(mtx, times = 1L, clockwise = TRUE)
}
\arguments{
\item{mtx}{a 2D matrix}

\item{times}{integer, how often to rotate in 90 degree steps. Example: pass \code{3L} to rotate \code{270} degrees.}

\item{clockwise}{logical, whether to rotate clockwise.}
}
\description{
Rotate 2D matrix clockwise in 90 degree steps.
}
\keyword{internal}
