% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{scan.mrtrix.file}
\alias{scan.mrtrix.file}
\title{Scan an MRtrix TCK or TSF file.}
\usage{
scan.mrtrix.file(filepath, want, chunk_values = 4e+06)
}
\arguments{
\item{filepath}{character string, path to the file.}

\item{want}{character vector, the values to compute, see
\code{scan.dti.tract.file}.}

\item{chunk_values}{integer, number of payload values to read per chunk.}
}
\value{
named list with entries \code{count} and \code{bbox}.
}
\description{
Scan an MRtrix TCK or TSF file.
}
\keyword{internal}
