% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{scan.trk.file}
\alias{scan.trk.file}
\title{Scan a TRK file.}
\usage{
scan.trk.file(filepath, want)
}
\arguments{
\item{filepath}{character string, path to the file.}

\item{want}{character vector, the values to compute, see
\code{scan.dti.tract.file}.}
}
\value{
named list with entries \code{count} and \code{bbox}.
}
\description{
Scan a TRK file.
}
\keyword{internal}
