% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_transform.R
\name{scann}
\alias{scann}
\title{Scan exactly n values from source string.}
\usage{
scann(cstring, num = 1L, what = integer(), line_number = NULL)
}
\arguments{
\item{cstring}{the input character string}

\item{num}{integer, the number of expected resulting items.}

\item{line_number}{optional integer, the line number (if the string represents a line from a text file). Will be printed in error message, if any.}
}
\value{
vector of type integer or double
}
\description{
Scan exactly n values from source string.
}
\keyword{internal}
