% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{stl.format.file.is.ascii}
\alias{stl.format.file.is.ascii}
\title{Guess whether a mesh file in STL format is the ASCII or the binary version.}
\usage{
stl.format.file.is.ascii(filepath)
}
\arguments{
\item{filepath}{full path to surface mesh file in STL format.}
}
\description{
Guess whether a mesh file in STL format is the ASCII or the binary version.
}
\keyword{internal}
