% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{[.fs.tracts}
\alias{[.fs.tracts}
\title{Subset an fs.tracts instance.}
\usage{
\method{[}{fs.tracts}(x, i)
}
\arguments{
\item{x}{an \code{fs.tracts} instance.}

\item{i}{index vector (integer, numeric or logical), as usual in R.}
}
\value{
a new \code{fs.tracts} instance containing the selected tracts.
}
\description{
Subset an fs.tracts instance.
}
\examples{
\dontrun{
tck <- read.dti.tck("brain.tck");
first_ten <- tck$tracks[1:10];
}

}
