% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_tracts.R
\name{[[.fs.tracts}
\alias{[[.fs.tracts}
\title{Access a single tract of an fs.tracts instance.}
\usage{
\method{[[}{fs.tracts}(x, i)
}
\arguments{
\item{x}{an \code{fs.tracts} instance.}

\item{i}{positive integer, the index of the tract to retrieve.}
}
\value{
the tract, see the description.
}
\description{
For TCK data (see \code{\link{read.dti.tck}}), a tract is an n x 3
numeric matrix of coordinates. For TRK data (see \code{\link{read.dti.trk}}), a
tract is a named list with the entries \code{coords} (n x 3 matrix),
\code{num_points} (integer), \code{scalars} (n x n_scalars matrix or NULL)
and \code{properties} (numeric vector or NULL).

This returns a copy of the requested tract, the data is stored in a single
matrix internally.
}
\examples{
\dontrun{
tck <- read.dti.tck("brain.tck");
first_tract_coords <- tck$tracks[[1]];
}

}
