% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_convert.R
\name{transform.descriptor.path}
\alias{transform.descriptor.path}
\title{Get the file path recorded in a volume descriptor.}
\usage{
\method{transform}{descriptor.path}(descriptor)
}
\arguments{
\item{descriptor}{\code{NULL} or a volume descriptor, see \code{\link{volume.descriptor}}.}
}
\value{
\code{NULL} or character string.
}
\description{
Get the file path recorded in a volume descriptor.
}
\keyword{internal}
