% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_transform.R
\name{transform.values.text}
\alias{transform.values.text}
\title{Format numerical values for a transformation text file.}
\usage{
\method{transform}{values.text}(values)
}
\arguments{
\item{values}{numerical vector, the values to format.}
}
\value{
character string, the values separated by single spaces.
}
\description{
Format numerical values for a transformation text file.
}
\keyword{internal}
