% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms_convert.R
\name{transform2lps}
\alias{transform2lps}
\title{Convert a transformation to LPS world coordinates.}
\usage{
transform2lps(tf)
}
\arguments{
\item{tf}{an \code{fs.transform} instance whose matrix operates on world coordinates, i.e. \code{space_in} and
\code{space_out} are either 'lps' or 'ras'.}
}
\value{
an \code{fs.transform} instance whose matrix operates on LPS coordinates.
}
\description{
The reverse of \code{\link{transform2ras}}, for transformations that have to be expressed in
the world coordinates that ITK and the tools built on it use. A transformation that is already in LPS
coordinates is returned unchanged.
}
\examples{
tf <- read.fs.transform(system.file("extdata", "talairach.xfm",
  package = "freesurferformats", mustWork = TRUE
))
# The matrix changes, because the sign of the first two axes changes.
max(abs(transform2lps(tf)$matrix - tf$matrix)) > 0

}
\seealso{
Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.dat]{read.fs.transform.dat()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.itk]{read.fs.transform.itk()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=sm1to0]{sm1to0()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform2ras]{transform2ras()}},
\code{\link[=transform2voxel]{transform2voxel()}},
\code{\link[=transform2world]{transform2world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.dat]{write.fs.transform.dat()}},
\code{\link[=write.fs.transform.fslmat]{write.fs.transform.fslmat()}},
\code{\link[=write.fs.transform.itk]{write.fs.transform.itk()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}},
\code{\link[=write.fs.transform.xfm]{write.fs.transform.xfm()}}
}
\concept{header coordinate space}
