% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti_tracks_utils.R
\name{trk.track.iterator}
\alias{trk.track.iterator}
\title{Create an iterator over the tracts of a TRK file.}
\usage{
trk.track.iterator(filepath, skip_tracks = 0L, bbox = NULL)
}
\arguments{
\item{filepath}{character string, path to the file.}

\item{skip_tracks}{integer, the number of tracts to skip.}

\item{bbox}{numeric vector of length 6 or NULL, a bounding box, see
\code{read.dti.tck}.}
}
\value{
the iterator environment.
}
\description{
Create an iterator over the tracts of a TRK file.
}
\keyword{internal}
