% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{unpack.vertex.coords}
\alias{unpack.vertex.coords}
\title{Unpack vertex coords from 0-255 representation.}
\usage{
unpack.vertex.coords(coords_packed, mdl_header)
}
\arguments{
\item{coords_packed}{matrix of n x 3 integers in range 0..255, the packed coords.}

\item{mdl_header}{MDL header or named list, only the fields 'header$scale' and 'header$origin' are used.}
}
\description{
Unpack vertex coords from 0-255 representation.
}
\keyword{internal}
