% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{validate.bbox}
\alias{validate.bbox}
\title{Validate a bounding box argument.}
\usage{
validate.bbox(bbox)
}
\arguments{
\item{bbox}{numeric vector of length 6 or NULL.}
}
\value{
TRUE, invisibly. Stops if the box is invalid.
}
\description{
Validate a bounding box argument.
}
\keyword{internal}
