% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{vis.quakemodel}
\alias{vis.quakemodel}
\title{Visualize Quake or Quake II alias model.}
\usage{
vis.quakemodel(md2, texture_file = NULL, frame_idx = 1L)
}
\arguments{
\item{md2}{an MDL or MD2 model instance. Alternatively, a character string which will be interpreted as a file to load.}

\item{texture_file}{character string, path to model skin. Q2 MD2 only, Q1 models include the skin.}

\item{frame_idx}{integer, which frame to use for vertex positions. A model contains many vertex positions if it includes model animation data.}
}
\description{
Visualize Quake or Quake II alias model.
}
