% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.as.indices}
\alias{vtk.as.indices}
\title{Convert values that are used as indices to integers.}
\usage{
vtk.as.indices(values, filepath)
}
\arguments{
\item{values}{numeric vector, the values to convert.}

\item{filepath}{character string, the file the values were read from.}
}
\value{
integer vector.
}
\description{
Convert values that are used as indices to integers.
}
\keyword{internal}
