% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.next.line}
\alias{vtk.next.line}
\title{Read the next line of an ASCII VTK legacy file.}
\usage{
vtk.next.line(reader)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}
}
\value{
character string, or \code{NULL} at the end of the file.
}
\description{
Read the next line of an ASCII VTK legacy file.
}
\keyword{internal}
