% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.read.header}
\alias{vtk.read.header}
\title{Read the header of a VTK legacy file and set up the reader.}
\usage{
vtk.read.header(reader)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}
}
\value{
named list with the entries 'version' (character), 'encoding' ('ASCII'
or 'BINARY') and 'dataset' (character, the dataset type).
}
\description{
Read the header of a VTK legacy file and set up the reader.
}
\keyword{internal}
