% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.reader.bytes}
\alias{vtk.reader.bytes}
\title{Consume bytes from a VTK legacy reader.}
\usage{
vtk.reader.bytes(reader, num_bytes)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}

\item{num_bytes}{single non-negative integer, the number of bytes to consume.}
}
\value{
raw vector of length \code{num_bytes}.
}
\description{
Consume bytes from a VTK legacy reader.
}
\keyword{internal}
