% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.reader.close}
\alias{vtk.reader.close}
\title{Close the connection of a VTK legacy reader.}
\usage{
vtk.reader.close(reader)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Close the connection of a VTK legacy reader.
}
\keyword{internal}
