% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.reader.peek}
\alias{vtk.reader.peek}
\title{Read bytes from a VTK legacy reader without consuming them.}
\usage{
vtk.reader.peek(reader, num_bytes)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}

\item{num_bytes}{single non-negative integer, the number of bytes to peek at.}
}
\value{
raw vector, possibly shorter than \code{num_bytes} at the end of the file.
}
\description{
Read bytes from a VTK legacy reader without consuming them.
}
\keyword{internal}
