% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.section.values}
\alias{vtk.section.values}
\title{Read numeric values from a VTK legacy section.}
\usage{
vtk.section.values(reader, num_values, type_info)
}
\arguments{
\item{reader}{an environment as returned by \code{\link{vtk.reader.new}}.}

\item{num_values}{single non-negative integer, the number of values to read.}

\item{type_info}{named list, the result of \code{\link{vtk.data.type.info}}.}
}
\value{
numeric or integer vector of length \code{num_values}.
}
\description{
Read numeric values from a VTK legacy section.
}
\keyword{internal}
