% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtk_legacy.R
\name{vtk.split.line}
\alias{vtk.split.line}
\title{Split a line of a VTK legacy file into whitespace separated tokens.}
\usage{
vtk.split.line(line)
}
\arguments{
\item{line}{character string, the line to split.}
}
\value{
character vector, the tokens of the line.
}
\description{
Split a line of a VTK legacy file into whitespace separated tokens.
}
\keyword{internal}
