% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_surface.R
\name{vtk.write.surface.ascii}
\alias{vtk.write.surface.ascii}
\title{Write the sections of a triangular mesh in VTK ASCII format.}
\usage{
vtk.write.surface.ascii(con, vertex_coords, faces, version)
}
\arguments{
\item{con}{a connection opened for writing.}

\item{vertex_coords}{n x 3 matrix of doubles, the vertex coordinates.}

\item{faces}{n x 3 matrix of integers, the vertex indices of the faces,
already converted to zero-based indices.}

\item{version}{character string, either '4.2' or '5.1'.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Write the sections of a triangular mesh in VTK ASCII format.
}
\keyword{internal}
