% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_analyze.R
\name{write.analyze.header.internal}
\alias{write.analyze.header.internal}
\title{Write the 348 byte ANALYZE 7.5 header to a connection.}
\usage{
write.analyze.header.internal(fh, analyzeheader)
}
\arguments{
\item{fh}{connection to write to.}

\item{analyzeheader}{named list, a valid ANALYZE 7.5 header.}
}
\value{
invisible \code{NULL}, called for the side effect of writing to the connection.
}
\description{
Write the 348 byte ANALYZE 7.5 header to a connection.
}
\keyword{internal}
