% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti2.R
\name{write.char.zero.fill}
\alias{write.char.zero.fill}
\title{Write given dstring to binary file, fill with zeroes to reach a total length of 'nchars'.}
\usage{
write.char.zero.fill(dstring, filehandle, nchars)
}
\description{
Write given dstring to binary file, fill with zeroes to reach a total length of 'nchars'.
}
\keyword{internal}
