% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_annot.R
\name{write.fs.colormap}
\alias{write.fs.colormap}
\title{Write colormap file in FreeSurfer ACII LUT format.}
\usage{
write.fs.colormap(filepath, annot)
}
\arguments{
\item{filepath, }{string. Full path to the output colormap file.}

\item{annot}{An annotation, as returned by [read.fs.annot].}
}
\value{
the data.frame that was written to the LUT file.
}
\description{
Write the colormap of an annotation to a text file in FreeSurfer ASCII colormap lookup table (LUT) format.
}
\seealso{
Other atlas functions: \code{\link{read.fs.annot}}
}
\concept{atlas functions}
