% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_annot.R
\name{write.fs.colortable.from.annot}
\alias{write.fs.colortable.from.annot}
\title{Write colortable file in FreeSurfer ASCII LUT format from annotation.}
\usage{
write.fs.colortable.from.annot(filepath, annot)
}
\arguments{
\item{filepath, }{string. Full path to the output colormap file.}

\item{annot}{An annotation, as returned by [read.fs.annot()]. If you want to assign specific indices, you can add a column named 'struct_index' to the data.frame \code{annot$colortable_df}. If there is no such columns, the indices will be created automatically in the order of the regions, starting at zero.}
}
\value{
the data.frame that was written to the LUT file.
}
\description{
Write the colortable of an annotation to a text file in FreeSurfer ASCII colortable lookup table (LUT) format. An example file is `FREESURFER_HOME/FreeSurferColorLUT.txt`.
}
\seealso{
Other atlas functions: 
\code{\link{read.fs.annot}()},
\code{\link{read.fs.colortable}()},
\code{\link{write.fs.colortable}()}

Other colorLUT functions: 
\code{\link{read.fs.colortable}()},
\code{\link{write.fs.colortable}()}
}
\concept{atlas functions}
\concept{colorLUT functions}
