% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_curv.R
\name{write.fs.morph}
\alias{write.fs.morph}
\title{Write morphometry data in a format derived from the given file name.}
\usage{
write.fs.morph(filepath, data, format = "auto", ...)
}
\arguments{
\item{filepath}{string. The full file name. The format to use will be derived from the last characters, the suffix. Supported suffixes are "mgh" for MGH format, "mgz" for MGZ format, "smp" for Brainvoyager SMP format, "nii" or "nii.gz" for NIFTI v1 format, "gii" or "gii.gz" for GIFTI format, everything else will be treated as curv format.}

\item{data}{numerical vector. The data to write.}

\item{format}{character string, the format to use. One of c("auto", "mgh", "mgz", "curv", "n1", "ni2", "gii"). The default setting "auto" will determine the format from the file extension.}

\item{...}{additional parameters to pass to the respective writer function.}
}
\value{
character string. The format that was used to write the data. One of c("auto", "mgh", "mgz", "curv", "ni1", "ni2", "gii").
}
\description{
Given data and a morphometry file name, derive the proper format from the file extension and write the file.
}
\note{
A file name that is one of the standard CIFTI-2 names (e.g. \code{.dscalar.nii} or
\code{.dlabel.nii}) is an error: such a file has to contain the CIFTI XML metadata, and
a NIFTI file with that name is read as a CIFTI file (and refused) by this package
and by other software. Use \code{\link{write.fs.morph.cifti}} to write morphometry
data to a CIFTI-2 file, or \code{\link{write.cifti}} for the other CIFTI-2 types.
}
\seealso{
Other morphometry functions:
\code{\link[=fs.get.morph.file.ext.for.format]{fs.get.morph.file.ext.for.format()}},
\code{\link[=fs.get.morph.file.format.from.filename]{fs.get.morph.file.format.from.filename()}},
\code{\link[=read.fs.curv]{read.fs.curv()}},
\code{\link[=read.fs.mgh]{read.fs.mgh()}},
\code{\link[=read.fs.morph]{read.fs.morph()}},
\code{\link[=read.fs.morph.gii]{read.fs.morph.gii()}},
\code{\link[=read.fs.volume]{read.fs.volume()}},
\code{\link[=read.fs.weight]{read.fs.weight()}},
\code{\link[=write.fs.curv]{write.fs.curv()}},
\code{\link[=write.fs.label.gii]{write.fs.label.gii()}},
\code{\link[=write.fs.mgh]{write.fs.mgh()}},
\code{\link[=write.fs.morph.asc]{write.fs.morph.asc()}},
\code{\link[=write.fs.morph.gii]{write.fs.morph.gii()}},
\code{\link[=write.fs.morph.ni1]{write.fs.morph.ni1()}},
\code{\link[=write.fs.morph.ni2]{write.fs.morph.ni2()}},
\code{\link[=write.fs.morph.smp]{write.fs.morph.smp()}},
\code{\link[=write.fs.morph.txt]{write.fs.morph.txt()}},
\code{\link[=write.fs.weight]{write.fs.weight()}},
\code{\link[=write.fs.weight.asc]{write.fs.weight.asc()}}
}
\concept{morphometry functions}
