% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_surface.R
\name{write.fs.surface.byu}
\alias{write.fs.surface.byu}
\title{Write mesh to file in BYU ASCII format.}
\usage{
write.fs.surface.byu(filepath, vertex_coords, faces)
}
\arguments{
\item{filepath}{string. Full path to the output surface file, should end with '.byu', but that is not enforced.}

\item{vertex_coords}{n x 3 matrix of doubles. Each row defined the x,y,z coords for a vertex.}

\item{faces}{n x 3 matrix of integers. Each row defined the 3 vertex indices that make up the face. WARNING: Vertex indices should be given in R-style, i.e., the index of the first vertex is 1. However, they will be written in FreeSurfer style, i.e., all indices will have 1 substracted, so that the index of the first vertex will be zero.}
}
\value{
string the format that was written. One of "tris" or "quads". Currently only triangular meshes are supported, so always 'tris'.
}
\description{
Write mesh to file in BYU ASCII format.
}
\note{
This is a fixed field length ASCII format. Keep in mind that the BYU format expects the coordinates to be in the cube -1 to +1 on all three axes.
}
\examples{
\dontrun{
# Read a surface from a file:
surface_file <- system.file("extdata", "lh.tinysurface",
  package = "freesurferformats", mustWork = TRUE
)
mesh <- read.fs.surface(surface_file)
# Now save it:
write.fs.surface.byu(tempfile(fileext = ".byu"), mesh$vertices, mesh$faces)
}

}
\seealso{
Other mesh functions:
\code{\link[=faces.quad.to.tris]{faces.quad.to.tris()}},
\code{\link[=read.fs.surface]{read.fs.surface()}},
\code{\link[=read.fs.surface.asc]{read.fs.surface.asc()}},
\code{\link[=read.fs.surface.bvsrf]{read.fs.surface.bvsrf()}},
\code{\link[=read.fs.surface.geo]{read.fs.surface.geo()}},
\code{\link[=read.fs.surface.gii]{read.fs.surface.gii()}},
\code{\link[=read.fs.surface.ico]{read.fs.surface.ico()}},
\code{\link[=read.fs.surface.obj]{read.fs.surface.obj()}},
\code{\link[=read.fs.surface.off]{read.fs.surface.off()}},
\code{\link[=read.fs.surface.ply]{read.fs.surface.ply()}},
\code{\link[=read.fs.surface.vtk]{read.fs.surface.vtk()}},
\code{\link[=read.mesh.brainvoyager]{read.mesh.brainvoyager()}},
\code{\link[=read_nisurface]{read_nisurface()}},
\code{\link[=read_nisurfacefile]{read_nisurfacefile()}},
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.asc]{write.fs.surface.asc()}},
\code{\link[=write.fs.surface.gii]{write.fs.surface.gii()}},
\code{\link[=write.fs.surface.mz3]{write.fs.surface.mz3()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh functions}
