% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_transform.R
\name{write.fs.transform.fslmat}
\alias{write.fs.transform.fslmat}
\title{Write a transformation matrix in FSL format.}
\usage{
write.fs.transform.fslmat(tf, filepath)
}
\arguments{
\item{tf}{an \code{fs.transform} instance whose matrix maps voxel coordinates to voxel coordinates.}

\item{filepath}{character string, the full path of the file to write.}
}
\value{
the \code{fs.transform} instance \code{tf}, invisibly.
}
\description{
Write a 4x4 matrix as an FSL matrix file, i.e. as the plain text file that FSL's \code{flirt} writes
with the \code{-omat} option and that FSL, MRtrix3 and FreeSurfer read as the registration between two images. The
matrix must map voxel coordinates to voxel coordinates (\code{space_in} and \code{space_out} are 'voxel'), because that
is what an FSL matrix stores: it relates the voxel grid of the image given to \verb{flirt -in} to the voxel grid of
the image given to \code{flirt -ref}, and it does not record which images those were. Use
\code{\link{transform2voxel}} to convert a transformation in world coordinates into one that can be
written.
}
\examples{
tf_file <- system.file("extdata", "talairach.lta", package = "freesurferformats", mustWork = TRUE)
out_file <- tempfile(fileext = ".mat")
write.fs.transform.fslmat(read.fs.transform(tf_file), out_file)
readLines(out_file)
unlink(out_file)

}
\seealso{
Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.dat]{read.fs.transform.dat()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.itk]{read.fs.transform.itk()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=sm1to0]{sm1to0()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform2lps]{transform2lps()}},
\code{\link[=transform2ras]{transform2ras()}},
\code{\link[=transform2voxel]{transform2voxel()}},
\code{\link[=transform2world]{transform2world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.dat]{write.fs.transform.dat()}},
\code{\link[=write.fs.transform.itk]{write.fs.transform.itk()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}},
\code{\link[=write.fs.transform.xfm]{write.fs.transform.xfm()}}
}
\concept{header coordinate space}
