% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_transform.R
\name{write.fs.transform.xfm}
\alias{write.fs.transform.xfm}
\title{Write an MNI transform (xfm) file.}
\usage{
write.fs.transform.xfm(tf, filepath, type = "Linear")
}
\arguments{
\item{tf}{an \code{fs.transform} instance whose matrix maps RAS coordinates to RAS coordinates.}

\item{filepath}{character string, the full path of the file to write.}

\item{type}{character string, the value of the \code{Transform_Type} entry, 'Linear' by default.}
}
\value{
the \code{fs.transform} instance \code{tf}, invisibly.
}
\description{
Write a transformation in the FreeSurfer xfm format, the format of \code{talairach.xfm} files, which
store the transformation from the RAS space of a subject to the RAS space of an MNI or Talairach template.
Only the first three rows of the matrix are stored, so the transformation must be affine, and it must map RAS
coordinates to RAS coordinates.
}
\examples{
xfm_file <- system.file("extdata", "talairach.xfm", package = "freesurferformats", mustWork = TRUE)
tf <- read.fs.transform(xfm_file)
out_file <- tempfile(fileext = ".xfm")
write.fs.transform.xfm(tf, out_file)
max(abs(read.fs.transform(out_file)$matrix - tf$matrix)) # 0
unlink(out_file)

}
\seealso{
Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.dat]{read.fs.transform.dat()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.itk]{read.fs.transform.itk()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=sm1to0]{sm1to0()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform.to.lps]{transform.to.lps()}},
\code{\link[=transform.to.ras]{transform.to.ras()}},
\code{\link[=transform.to.voxel]{transform.to.voxel()}},
\code{\link[=transform.to.world]{transform.to.world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.dat]{write.fs.transform.dat()}},
\code{\link[=write.fs.transform.fslmat]{write.fs.transform.fslmat()}},
\code{\link[=write.fs.transform.itk]{write.fs.transform.itk()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}}
}
\concept{header coordinate space}
