% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti1.R
\name{write.nifti1.data.internal}
\alias{write.nifti1.data.internal}
\title{Write the voxel data of a NIFTI v1 file to a connection.}
\usage{
write.nifti1.data.internal(fh, niidata, niiheader)
}
\arguments{
\item{fh}{connection to write to.}

\item{niidata}{array of numeric data, or \code{NULL} to write no data at all.}

\item{niiheader}{named list, a valid NIFTI v1 header that describes the data.}
}
\value{
the data that was written, after conversion to the data type stated in the header, or \code{NULL} if \code{niidata}
was \code{NULL}.
}
\description{
Write the voxel data of a NIFTI v1 file to a connection.
}
\keyword{internal}
