% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti1.R
\name{write.nifti1.header.internal}
\alias{write.nifti1.header.internal}
\title{Write the 348 byte NIFTI v1 header (and the padding up to the data offset) to a connection.}
\usage{
write.nifti1.header.internal(fh, niiheader)
}
\arguments{
\item{fh}{connection to write to.}

\item{niiheader}{named list, a valid NIFTI v1 header.}
}
\value{
invisible \code{NULL}, called for the side effect of writing to the connection.
}
\description{
Write the 348 byte NIFTI v1 header (and the padding up to the data offset) to a connection.
}
\keyword{internal}
