% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvoyager.R
\name{write.smp.brainvoyager}
\alias{write.smp.brainvoyager}
\title{Write a brainvoyager SMP file.}
\usage{
write.smp.brainvoyager(filepath, bvsmp, smp_version = 3L)
}
\arguments{
\item{filepath}{character string, the output file}

\item{bvsmp}{a named list, as returned by \code{\link{read.smp.brainvoyager}}.}

\item{smp_version}{integer, the SMP file format version to use when writing. Only v3 is supported atm.}
}
\description{
Write a brainvoyager SMP file, which contains one or more vertex-wise data maps (stats or morpometry data).
}
\seealso{
\code{\link{write.fs.morph.smp}}
}
