% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvoyager.R
\name{write.smp.brainvoyager.v2}
\alias{write.smp.brainvoyager.v2}
\title{Write a brainvoyager v2 SMP file.}
\usage{
write.smp.brainvoyager.v2(filepath, bvsmp)
}
\arguments{
\item{filepath}{character string, the output file}

\item{bvsmp}{bvsmp instance, a named list as returned by \code{\link{read.smp.brainvoyager}}.}
}
\description{
Write a brainvoyager v2 SMP file.
}
\note{
Called by \code{\link{write.smp.brainvoyager}}.

The map_type and num_lags of the first vertex map will be used for the top header,i.e., for all maps. The v2 format does not support per-map settings for these values. Also min_alg, max_lag and cc_overlay are ignored.
}
\keyword{internal}
