% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvoyager.R
\name{write.smp.brainvoyager.v3}
\alias{write.smp.brainvoyager.v3}
\title{Write a brainvoyager v3 SMP file.}
\usage{
write.smp.brainvoyager.v3(filepath, bvsmp)
}
\arguments{
\item{filepath}{character string, the output file}

\item{bvsmp}{a named list, as returned by \code{\link{read.smp.brainvoyager}}.}
}
\description{
Write a brainvoyager v3 SMP file.
}
\note{
Called by \code{\link{write.smp.brainvoyager}}.
}
\keyword{internal}
