% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainvoyager.R
\name{write.smp.brainvoyager.v3or4or5}
\alias{write.smp.brainvoyager.v3or4or5}
\title{Write a brainvoyager v3, v4 or v5 SMP file.}
\usage{
write.smp.brainvoyager.v3or4or5(filepath, bvsmp, smp_version)
}
\arguments{
\item{filepath}{character string, the output file}

\item{bvsmp}{bvsmp instance, a named list as returned by \code{\link{read.smp.brainvoyager}}.}

\item{smp_version}{integer, the SMP file format version to use when writing. Versions 2 to 5 are supported, but only versions 2 and 3 have been tested properly. Please report any problems you encounter. When converting between file versions (e.g., loading a v2 file and saving the result as a v5 file), some required fields may be missing, and for those without a default value according to the official spec, you will have to manually add the value you want in the bvsmp object before writing.}
}
\description{
Write a brainvoyager v3, v4 or v5 SMP file.
}
\note{
Called by \code{\link{write.smp.brainvoyager}}.
}
\keyword{internal}
