% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_xml_tools.R
\name{xml_node_gifti_label_table}
\alias{xml_node_gifti_label_table}
\title{Create XML GIFTI LabelTable node.}
\usage{
xml_node_gifti_label_table(attributes = list())
}
\value{
XML node from xml2
}
\description{
Create XML GIFTI LabelTable node.
}
\keyword{internal}
