% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_xml_tools.R
\name{xml_node_gifti_label_table_from_annot}
\alias{xml_node_gifti_label_table_from_annot}
\title{Compute LabelTable node from annotation.}
\usage{
xml_node_gifti_label_table_from_annot(annot)
}
\arguments{
\item{annot}{an fs.annotation, the included data will be used to compute the LabelTable node}
}
\value{
XML tree from xml2, the LabelTable and its child nodes
}
\description{
Compute LabelTable node from annotation.
}
\keyword{internal}
