% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_xml_tools.R
\name{xml_node_gifti_MD}
\alias{xml_node_gifti_MD}
\title{Create XML GIFTI metadata node.}
\usage{
xml_node_gifti_MD(name, value, as_cdata = TRUE)
}
\arguments{
\item{name}{character string, the metadata name}

\item{value}{character string, the metadata value}

\item{as_cdata}{logical, whether to wrap the value in cdata tags}
}
\value{
XML tree from xml2
}
\description{
Create XML GIFTI metadata node.
}
\keyword{internal}
