% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.axis.labels}
\alias{cifti.axis.labels}
\title{Get axis labels for a matrix dimension of a CIFTI-2 file.}
\usage{
cifti.axis.labels(cii, dim = 0L)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension to get the brain models for. CIFTI-2
files have two dimensions, so this is usually 0 (Workbench calls it the ROW
dimension) or 1 (the COLUMN dimension). See
\code{read.cifti.header} for details on the dimensions.}
}
\value{
character vector of length \code{cii$matrix$dim_sizes[dim + 1]}, the labels,
or \code{NULL} if no labels are known for this kind of dimension.
}
\description{
Get a label for each index of one matrix dimension of a CIFTI-2
file, e.g. the map names of a \code{.dscalar}, the parcel names of a \code{.pconn} or
identifiers for the grayordinates of a dense file. This is what
\code{\link{read.cifti}} uses to name the dimensions of the data array.
}
\examples{
cifti_file <- system.file("extdata", "cifti", "tiny.ptseries.nii", package = "freesurferformats")
cii <- read.cifti.header(cifti_file)
cifti.axis.labels(cii, dim = 1L)

}
\seealso{
Other cifti functions:
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}}
}
\concept{cifti functions}
