% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_axes.R
\name{cifti.axis.parcels}
\alias{cifti.axis.parcels}
\title{Create a CIFTI-2 axis for parcels.}
\usage{
cifti.axis.parcels(parcels, surfaces = NULL, volume = NULL)
}
\arguments{
\item{parcels}{list of parcels, as created by \code{\link{cifti.parcel}}.}

\item{surfaces}{named integer vector or \code{NULL}, the number of vertices of the complete
surfaces the parcels refer to, see \code{\link{cifti.axis.brain.models}}. Unlike
Connectome Workbench, which writes them, this is optional here as well.}

\item{volume}{a volume, see \code{\link{cifti.volume}}, for parcels that consist of
volume voxels, or \code{NULL}.}
}
\value{
a named list with the entries 'type', 'parcels', 'surfaces' and 'volumes', an
axis to be passed to \code{\link{cifti.header.from.axes}} or
\code{\link{write.cifti}}.
}
\description{
A parcel axis describes the parcels of a parcellated matrix dimension (the
mapping of \code{.pscalar}, \code{.ptseries}, \code{.pconn} and the second dimension of \code{.pdconn}).
As for brain models, the index of a parcel is its position in the list.
}
\examples{
axis <- cifti.axis.parcels(list(
  cifti.parcel("PARCEL_A", list(CORTEX_LEFT = 0:2)),
  cifti.parcel("PARCEL_B", list(CORTEX_LEFT = 3:5))))
length(axis$parcels)

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.from.template]{cifti.axis.from.template()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels.from.annot]{cifti.axis.parcels.from.annot()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.axis.series]{cifti.axis.series()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.file.type.for.axes]{cifti.file.type.for.axes()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=cifti.volume]{cifti.volume()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=print.fs.connectome]{print.fs.connectome()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=read.fs.connectome.cifti]{read.fs.connectome.cifti()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.connectome.cifti]{write.fs.connectome.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellated.cifti]{write.fs.parcellated.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
