% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_axes.R
\name{cifti.axis.series}
\alias{cifti.axis.series}
\title{Create a CIFTI-2 axis for a series.}
\usage{
cifti.axis.series(
  number_of_series_points,
  start = 0,
  step = 1,
  exponent = 0L,
  unit = "SECOND"
)
}
\arguments{
\item{number_of_series_points}{integer, the number of samples, which must match the
size of that matrix dimension of the data.}

\item{start}{numeric, the value of the first sample.}

\item{step}{numeric, the difference between consecutive samples.}

\item{exponent}{integer, the power of ten the values are in (e.g. -3 for milliseconds).}

\item{unit}{character string, one of 'SECOND', 'HERTZ', 'METER' or 'RADIAN'.}
}
\value{
a named list with the entries 'type' and 'series', an axis to be passed to
\code{\link{cifti.header.from.axes}} or \code{\link{write.cifti}}.
}
\description{
A series axis describes a dimension that holds an ordered sequence of
samples, usually the time points of a \code{.dtseries} or \code{.ptseries} file. The value of
series index \code{i} (0-based) is \code{(start + i * step) * 10^exponent}, in the unit that
\code{unit} names.
}
\examples{
axis <- cifti.axis.series(10L, start = 0, step = 0.72, unit = "SECOND")
axis$series$number_of_series_points

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.from.template]{cifti.axis.from.template()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels]{cifti.axis.parcels()}},
\code{\link[=cifti.axis.parcels.from.annot]{cifti.axis.parcels.from.annot()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.file.type.for.axes]{cifti.file.type.for.axes()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=cifti.volume]{cifti.volume()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=print.fs.connectome]{print.fs.connectome()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=read.fs.connectome.cifti]{read.fs.connectome.cifti()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.connectome.cifti]{write.fs.connectome.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellated.cifti]{write.fs.parcellated.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
