% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{cifti.file.type.for.axes}
\alias{cifti.file.type.for.axes}
\title{Get the standard CIFTI-2 file type for a set of axes.}
\usage{
cifti.file.type.for.axes(axes)
}
\arguments{
\item{axes}{list of axes, see \code{\link{cifti.header.from.axes}}.}
}
\value{
a one row data.frame, see \code{\link{cifti.file.types}}, or a row with
\code{intent_code = 3000} ('ConnUnknown') and empty file name extension for a combination
that the standard does not define.
}
\description{
Look up the file type that the format defines for the combination of matrix
index types of the axes, see \code{\link{cifti.file.types}}.
}
\examples{
axis_brain <- cifti.axis.brain.models(list(cifti.brain.model.surface("lh", 10L)))
type <- cifti.file.type.for.axes(list(cifti.axis.scalars("m"), axis_brain))
type$extension

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.from.template]{cifti.axis.from.template()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels]{cifti.axis.parcels()}},
\code{\link[=cifti.axis.parcels.from.annot]{cifti.axis.parcels.from.annot()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.axis.series]{cifti.axis.series()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=cifti.volume]{cifti.volume()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=print.fs.connectome]{print.fs.connectome()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=read.cifti.rows]{read.cifti.rows()}},
\code{\link[=read.fs.connectome.cifti]{read.fs.connectome.cifti()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.connectome.cifti]{write.fs.connectome.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellated.cifti]{write.fs.parcellated.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
