% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{cifti.grayordinates}
\alias{cifti.grayordinates}
\title{Get the brainordinate table of a CIFTI-2 file.}
\usage{
cifti.grayordinates(cii, dim = 0L)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension to get the brain models for. CIFTI-2
files have two dimensions, so this is usually 0 (Workbench calls it the ROW
dimension) or 1 (the COLUMN dimension). See
\code{read.cifti.header} for details on the dimensions.}
}
\value{
a data.frame with one row per index of the matrix dimension and the
columns 'index' (integer, 0-based matrix index), 'structure' (character
string, the brain structure as spelled in the file), 'structure_short'
(without the \code{CIFTI_STRUCTURE_} prefix), 'model_type' ('SURFACE' or
'VOXELS'), 'vertex_index' (integer, 0-based index of the vertex in the
surface for surface models, \code{NA} for volume models) and 'i', 'j', 'k'
(integer, 0-based voxel indices for volume models, \code{NA} for surface models).
}
\description{
Get the mapping from the matrix indices of one dimension to the
individual surface vertices and volume voxels a CIFTI-2 file contains. This
is the per-index version of the table returned by
\code{\link{cifti.structures}}, which reports index ranges instead of single
indices. It is useful to find out which vertex or voxel a data row belongs
to, or to find the data index of a vertex or voxel.

Note that the same vertex or voxel can only appear once per dimension (the
matrix has one entry per brainordinate), but a structure can be split over
several brain model entries, and the vertices of a grayordinates file are a
subset of the vertices of the surface it refers to.
}
\examples{
cifti_file <- system.file("extdata", "cifti", "tiny.dscalar.nii", package = "freesurferformats")
cii <- read.cifti.header(cifti_file)
grayordinates <- cifti.grayordinates(cii, dim = 1L)
head(grayordinates)

}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.brain.models]{cifti.axis.brain.models()}},
\code{\link[=cifti.axis.from.template]{cifti.axis.from.template()}},
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.axis.parcels]{cifti.axis.parcels()}},
\code{\link[=cifti.axis.parcels.from.annot]{cifti.axis.parcels.from.annot()}},
\code{\link[=cifti.axis.scalars]{cifti.axis.scalars()}},
\code{\link[=cifti.axis.series]{cifti.axis.series()}},
\code{\link[=cifti.brain.model.surface]{cifti.brain.model.surface()}},
\code{\link[=cifti.brain.model.volume]{cifti.brain.model.volume()}},
\code{\link[=cifti.dim.labels]{cifti.dim.labels()}},
\code{\link[=cifti.file.type.for.axes]{cifti.file.type.for.axes()}},
\code{\link[=cifti.header.from.axes]{cifti.header.from.axes()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcel]{cifti.parcel()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=cifti.volume]{cifti.volume()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=print.fs.cifti.data]{print.fs.cifti.data()}},
\code{\link[=print.fs.connectome]{print.fs.connectome()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}},
\code{\link[=read.fs.connectome.cifti]{read.fs.connectome.cifti()}},
\code{\link[=write.cifti]{write.cifti()}},
\code{\link[=write.fs.connectome.cifti]{write.fs.connectome.cifti()}},
\code{\link[=write.fs.morph.cifti]{write.fs.morph.cifti()}},
\code{\link[=write.fs.parcellated.cifti]{write.fs.parcellated.cifti()}},
\code{\link[=write.fs.parcellation.cifti]{write.fs.parcellation.cifti()}},
\code{\link[=write.fs.series.cifti]{write.fs.series.cifti()}}
}
\concept{cifti functions}
