% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.parcels}
\alias{cifti.parcels}
\title{Get the parcel table of a CIFTI-2 file.}
\usage{
cifti.parcels(cii, dim = 0L)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension to get the brain models for. CIFTI-2
files have two dimensions, so this is usually 0 (Workbench calls it the ROW
dimension) or 1 (the COLUMN dimension). See
\code{read.cifti.header} for details on the dimensions.}
}
\value{
a data.frame with one row per parcel and the columns 'index' (integer,
0-based parcel index), 'name' (character string, the parcel name), 'num_vertices'
(integer, the total number of surface vertices in this parcel) and
'num_voxels' (integer, the number of volume voxels in this parcel). The
result is ordered by parcel index.
}
\description{
Get the parcels for one matrix dimension of a parcellated
CIFTI-2 file. A parcel is a named set of surface vertices and/or volume
voxels; its index is its position in the list (CIFTI-2 parcels have no index
attribute). The vertex lists themselves are in the field 'vertices' of the
\code{fs.cifti} object, see \code{\link{read.cifti.header}}.
}
\examples{
cifti_file <- system.file("extdata", "cifti", "tiny.ptseries.nii", package = "freesurferformats")
cii <- read.cifti.header(cifti_file)
cifti.parcels(cii, dim = 1L)

}
\seealso{
Other cifti functions:
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}}
}
\concept{cifti functions}
