% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{cifti.series.info}
\alias{cifti.series.info}
\title{Get the series information of a CIFTI-2 file.}
\usage{
cifti.series.info(cii, dim = 0L)
}
\arguments{
\item{cii}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{dim}{integer, the matrix dimension to get the brain models for. CIFTI-2
files have two dimensions, so this is usually 0 (Workbench calls it the ROW
dimension) or 1 (the COLUMN dimension). See
\code{read.cifti.header} for details on the dimensions.}
}
\value{
a named list with the entries 'number_of_series_points' (integer),
'start' (numeric), 'step' (numeric), 'exponent' (integer) and 'unit'
(character string, one of 'SECOND', 'HERTZ', 'METER', 'RADIAN'), or \code{NULL}
if the specified matrix dimension is not a series. To check whether a
dimension is a series, inspect the field 'type' of the corresponding entry
of \code{cii$matrix$indices_maps}.
}
\description{
Get the description of the series dimension of a CIFTI-2 file,
e.g. the time points of a \code{.dtseries} or \code{.ptseries} file. The time (or
other unit) of the series index \code{i} is
\code{(start + i * step) * 10^exponent}, with \code{i} starting at 0.
}
\examples{
cifti_file <- system.file("extdata", "cifti", "tiny.dtseries.nii", package = "freesurferformats")
cii <- read.cifti.header(cifti_file)
cifti.series.info(cii, dim = 0L)

}
\seealso{
Other cifti functions:
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}}
}
\concept{cifti functions}
