% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{faces.quad.to.tris}
\alias{faces.quad.to.tris}
\title{Convert quad faces to tris faces.}
\usage{
faces.quad.to.tris(quad_faces)
}
\arguments{
\item{quad_faces}{nx4 integer matrix, the indices of the vertices making up the *n* quad faces}
}
\value{
*2nx3* integer matrix, the indices of the vertices making up the *2n* tris faces
}
\description{
Convert quad faces to tris faces.
}
