% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_surface.R
\name{fixed.vec.format.int}
\alias{fixed.vec.format.int}
\title{Write fixed width integers to one or several lines.}
\usage{
fixed.vec.format.int(
  vdata,
  num_chars_per_entry,
  max_entries_per_line = NULL,
  align_right = TRUE
)
}
\arguments{
\item{vdata}{integer vector, the data}

\item{num_chars_per_entry}{field length of a single formatted integer in characters}

\item{max_entries_per_line}{integer, how many entries are allowed per line. Leave at NULL for no limit, which will return all in a single line.}

\item{align_right}{logical, whether to align the integers to the right. As you may have guessed, set to `FALSE` to align to the left.}
}
\value{
character string, the formatted data. May contain newlines.
}
\description{
Write fixed width integers to one or several lines.
}
