% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_curv.R
\name{fs.get.morph.file.ext.for.format}
\alias{fs.get.morph.file.ext.for.format}
\title{Determine morphometry file extension from format}
\usage{
fs.get.morph.file.ext.for.format(format)
}
\arguments{
\item{format, }{string. One of c("mgh", "mgz", "curv", "gii").}
}
\value{
file ext, string. The standard file extension for the format. (May be an empty string for some formats.)
}
\description{
Given a morphometry file format, derive the proper file extension.
}
\seealso{
Other morphometry functions: 
\code{\link{fs.get.morph.file.format.from.filename}()},
\code{\link{read.fs.curv}()},
\code{\link{read.fs.mgh}()},
\code{\link{read.fs.morph.gii}()},
\code{\link{read.fs.morph}()},
\code{\link{read.fs.volume}()},
\code{\link{read.fs.weight}()},
\code{\link{write.fs.curv}()},
\code{\link{write.fs.label.gii}()},
\code{\link{write.fs.mgh}()},
\code{\link{write.fs.morph.asc}()},
\code{\link{write.fs.morph.gii}()},
\code{\link{write.fs.morph.ni1}()},
\code{\link{write.fs.morph.ni2}()},
\code{\link{write.fs.morph.smp}()},
\code{\link{write.fs.morph.txt}()},
\code{\link{write.fs.morph}()},
\code{\link{write.fs.weight.asc}()},
\code{\link{write.fs.weight}()}
}
\concept{morphometry functions}
