% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_curv.R
\name{fs.get.morph.file.format.from.filename}
\alias{fs.get.morph.file.format.from.filename}
\title{Determine morphometry file format from filename}
\usage{
fs.get.morph.file.format.from.filename(filepath)
}
\arguments{
\item{filepath, }{string. A path to a file.}
}
\value{
format, string. The format, one of c("mgz", "mgh", "curv").
}
\description{
Given a morphometry file name, derive the proper file format, based on the suffix. Case is ignored, i.e., cast to lowercase before checks. If the filepath ends with "mgh", returns format "mgh". For suffix "mgz", returns "mgz" format. For all others, returns "curv" format.
}
\keyword{internal}
