% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_trk.R
\name{get.dti.trk.endianness}
\alias{get.dti.trk.endianness}
\title{Determine endianness of TRK file.}
\usage{
get.dti.trk.endianness(filepath)
}
\arguments{
\item{filepath}{character string, path to file in trk format.}
}
\value{
endina character string. one of 'little' or 'big'.
}
\description{
Determine endianness of TRK file.
}
\note{
This function checks endiannes via the header size field of the file header, which must be 1000 for TRK files when read with correct enianness. It will stop if the file is not in TRK format, i.e., if the field is not 1000 in any endianness.
}
\keyword{internal}
