% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{get.slice.orientation}
\alias{get.slice.orientation}
\title{Compute MGH orientation string and direction}
\usage{
get.slice.orientation(Mdc)
}
\arguments{
\item{Mdc}{numeric 3x3 matrix}
}
\value{
named list with entries: `orientation_string`: character string of length 3, one uppercase letter per axis. `direction_name`: slice direction, character string, one of 'sagittal', 'coronal', 'axial' or 'unknown'.
}
\description{
Compute MGH orientation string and direction
}
\keyword{internal}
