% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{is.fs.surface}
\alias{is.fs.surface}
\title{Check whether object is an fs.surface}
\usage{
is.fs.surface(x)
}
\arguments{
\item{x}{any `R` object}
}
\value{
TRUE if its argument is a brain surface (that is, has "fs.surface" amongst its classes) and FALSE otherwise.
}
\description{
Check whether object is an fs.surface
}
