% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.is.conformed}
\alias{mghheader.is.conformed}
\title{Determine whether an MGH volume is conformed.}
\usage{
mghheader.is.conformed(header)
}
\arguments{
\item{mgh_header}{Header of the mgh datastructure, as returned by \code{\link[freesurferformats]{read.fs.mgh}}.}
}
\value{
logical, whether the volume is *conformed*.
}
\description{
In the FreeSurfer sense, *conformed* means that the volume is in coronal primary slice direction, has dimensions 256x256x256 and a voxel size of 1 mm in all 3 directions. The slice direction can only be determined if the header contains RAS information, if it does not, the volume is not conformed.
}
