% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mghheader.is.ras.valid}
\alias{mghheader.is.ras.valid}
\title{Check whether header contains valid ras information}
\usage{
mghheader.is.ras.valid(header)
}
\arguments{
\item{header}{mgh header or `fs.volume` instance with header}
}
\value{
logical, whether header contains valid ras information
}
\description{
Check whether header contains valid ras information
}
\keyword{internal}
