% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_mgh.R
\name{mghheader.vox2ras}
\alias{mghheader.vox2ras}
\title{Compute vox2ras matrix from basic MGH header fields.}
\usage{
mghheader.vox2ras(header)
}
\arguments{
\item{header}{the MGH header}
}
\description{
This is also known as the 'scanner' vox2ras.
}
