% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{mni152reg}
\alias{mni152reg}
\title{Get fsaverage to MNI 152 transformation matrix.}
\usage{
mni152reg()
}
\description{
Get fsaverage to MNI 152 transformation matrix.
}
\note{
There are better ways to achieve this transformation than using this matrix, see Wu et al., 'Accurate nonlinear mapping between MNI volumetric and FreeSurfer surface coordinate system', Hum Brain Mapp. 2018 Sep; 39(9): 3793–3808. doi: 10.1002/hbm.24213.
}
\examples{
    coords_tf = doapply.transform.mtx(c(1.0, 1.0, 1.0), mni152reg());
    coords_tf;
    doapply.transform.mtx(coords_tf, solve(mni152reg()));

}
