% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti1.R
\name{ni1header.template}
\alias{ni1header.template}
\title{Create a template NIFTI v1 header. You will have to adapt it for your use case.}
\usage{
ni1header.template()
}
\value{
named list, the NIFTI v1 header. All fields are present and filled with values of a proper type. Whether or not they make sense is up to you, but you will most likely have to adapt at least the following fields to your data: `dim_raw`, `datatype`, `bitpix`.
}
\description{
Create a template NIFTI v1 header. You will have to adapt it for your use case.
}
\note{
Commonly used data type settings are: for signed integers datatype = `8L` and bitpix = `32L`; for floats datatype = `16L` and bitpix = `32L`. See the NIFTI v1 standard for more options.  You may want to call \code{\link{ni1header.for.data}} instead of this function.
}
\seealso{
\code{\link{ni1header.for.data}}
}
