% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_common.R
\name{nifti.datadim.to.dimfield}
\alias{nifti.datadim.to.dimfield}
\title{Compute NIFTI dim field for data dimension.}
\usage{
nifti.datadim.to.dimfield(datadim)
}
\arguments{
\item{datadim}{integer vector, the result of calling `dim` on your data. The length must be <= 7.}
}
\value{
NIFTI header `dim` field, an integer vector of length 8
}
\description{
Compute NIFTI dim field for data dimension.
}
\examples{
   nifti.datadim.to.dimfield(c(256, 256, 256));

}
\seealso{
Other NIFTI helper functions: 
\code{\link{nifti.datadim.from.dimfield}()}
}
\concept{NIFTI helper functions}
