% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.dtype.info}
\alias{nifti.dtype.info}
\title{Compute NIFTI v1 data type info from datatype and bitpix header field.}
\usage{
nifti.dtype.info(datatype, bitpix)
}
\arguments{
\item{datatype}{integer, the \code{datatype} NIFTI v1 header field}

\item{bitpix}{integer, the \code{bitpix} NIFTI v1 header field}
}
\value{
named list with entries: \code{mri_dtype}: the MRI data type, as used by FreeSurfer for MGH files, \code{r_dtype}: the R data type.
}
\description{
Compute NIFTI v1 data type info from datatype and bitpix header field.
}
\keyword{internal}
