% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti1.R
\name{nifti.file.uses.fshack}
\alias{nifti.file.uses.fshack}
\title{Determine whether NIFTI v1 file uses the FreeSurfer hack.}
\usage{
nifti.file.uses.fshack(filepath)
}
\arguments{
\item{filepath}{path to a NIFTI v1 file (single file version), which can contain the FreeSurfer hack.}
}
\value{
logical, whether the file header contains the FreeSurfer format hack. See \code{\link{nifti1.header}} for details.
}
\description{
Determine whether NIFTI v1 file uses the FreeSurfer hack.
}
