% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_nifti_hack.R
\name{nifti.header.fshack}
\alias{nifti.header.fshack}
\title{Read NIFTI v1 header from file with FreeSurfer hack.}
\usage{
nifti.header.fshack(filepath, little_endian = TRUE)
}
\arguments{
\item{filepath}{path to a NIFTI v1 file (single file version), which can contain the FreeSurfer hack.}

\item{little_endian}{internal logical, leave this alone. Endianness will be figured out automatically, messing with this parameter only hurts.}
}
\value{
named list with NIFTI 1 header fields.
}
\description{
Read NIFTI v1 header from file with FreeSurfer hack.
}
\note{
The FreeSurfer hack is a non-standard way to save long vectors in NIFTI v1 files.
}
