% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgh2nii.R
\name{nifti.space.info}
\alias{nifti.space.info}
\title{Compute NIFTI space unit info from xyzt_units header field.}
\usage{
nifti.space.info(xyzt_units)
}
\arguments{
\item{xyzt_units}{a single character, the `xyzt_units` NIFTI header field}
}
\value{
named list with entries: `code`: the NIFTI unit code as a decimal integer, `name`: character string, the unit name, `scaling`: float, the scaling factor for the unit, relative to the FreeSurfer space unit (`mm`).
}
\description{
Compute NIFTI space unit info from xyzt_units header field.
}
\keyword{internal}
