% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti_to_mgh.R
\name{nifti.time.info}
\alias{nifti.time.info}
\title{Compute NIFTI time unit info from xyzt_units header field.}
\usage{
nifti.time.info(xyzt_units)
}
\arguments{
\item{xyzt_units}{a single character, the `xyzt_units` NIFTI header field}
}
\value{
named list with entries: `code`: the NIFTI unit code as a decimal integer, `name`: character string, the unit name, `scaling`: float, the scaling factor for the unit, relative to the FreeSurfer time unit  (`ms`).
}
\description{
Compute NIFTI time unit info from xyzt_units header field.
}
\keyword{internal}
