% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{polygon.soup.to.indexed.mesh}
\alias{polygon.soup.to.indexed.mesh}
\title{Turn polygon soup into indexed mesh.}
\usage{
polygon.soup.to.indexed.mesh(faces_vertex_coords)
}
\arguments{
\item{face_vertex_coords}{numerical matrix with *n* rows and 3 columns, the vertex coordinates of the faces. Each row contains the x,y,z coordinates of a single vertex, and three consecutive vertex rows form a triangular face.}
}
\value{
a mesh, as an fs.surface instance
}
\description{
Turn polygon soup into indexed mesh.
}
\keyword{internal}
