% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_annot.R
\name{print.fs.annot}
\alias{print.fs.annot}
\title{Print description of a brain atlas or annotation.}
\usage{
\method{print}{fs.annot}(x, ...)
}
\arguments{
\item{x}{brain surface annotation or atlas with class `fs.annot`.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print description of a brain atlas or annotation.
}
