% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{print.fs.cifti.data}
\alias{print.fs.cifti.data}
\title{Print an fs.cifti.data instance.}
\usage{
\method{print}{fs.cifti.data}(x, ...)
}
\arguments{
\item{x}{an \code{fs.cifti.data} instance, see \code{\link{read.cifti}}.}

\item{...}{ignored.}
}
\value{
the instance, invisibly.
}
\description{
Print an fs.cifti.data instance.
}
\seealso{
Other cifti functions:
\code{\link[=cifti.axis.labels]{cifti.axis.labels()}},
\code{\link[=cifti.grayordinates]{cifti.grayordinates()}},
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structure.data]{cifti.structure.data()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=print.fs.cifti]{print.fs.cifti()}},
\code{\link[=read.cifti]{read.cifti()}},
\code{\link[=read.cifti.header]{read.cifti.header()}}
}
\concept{cifti functions}
