% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_header.R
\name{print.fs.cifti}
\alias{print.fs.cifti}
\title{Print an fs.cifti instance.}
\usage{
\method{print}{fs.cifti}(x, ...)
}
\arguments{
\item{x}{an \code{fs.cifti} instance, see \code{\link{read.cifti.header}}.}

\item{...}{ignored.}
}
\value{
the instance, invisibly.
}
\description{
Print an fs.cifti instance.
}
\seealso{
Other cifti functions:
\code{\link[=cifti.label.table]{cifti.label.table()}},
\code{\link[=cifti.parcels]{cifti.parcels()}},
\code{\link[=cifti.series.info]{cifti.series.info()}},
\code{\link[=cifti.structures]{cifti.structures()}},
\code{\link[=read.cifti.header]{read.cifti.header()}}
}
\concept{cifti functions}
