% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_label.R
\name{print.fs.label}
\alias{print.fs.label}
\title{Print description of a brain surface label.}
\usage{
\method{print}{fs.label}(x, ...)
}
\arguments{
\item{x}{brain surface label with class `fs.annot`.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print description of a brain surface label.
}
