% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_trk.R
\name{read.dti.trk}
\alias{read.dti.trk}
\title{Read fiber tracks from Diffusion Toolkit in trk format.}
\usage{
read.dti.trk(filepath)
}
\arguments{
\item{filepath}{character string, path to file in trk format.}
}
\value{
named list, the parsed file data.
}
\description{
Read fiber tracks from Diffusion Toolkit in trk format.
}
\examples{
\dontrun{
trk = read.dti.trk("~/simple.trk");
}

}
