% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dti_tcktsf.R
\name{read.dti.tsf}
\alias{read.dti.tsf}
\title{Read DTI tracking per-coord data from file in MRtrix 'TSF' format.}
\usage{
read.dti.tsf(filepath)
}
\arguments{
\item{filepath}{character string, path to the \code{TSF} file to read.}
}
\value{
named list with entries 'header' and 'scalars'. The scalar data are available in 2 representations: 'merged': a vector of all values (requires external knowledge on track borders), and 'scalar_list': organized into a list of vectors. Each vector represents the values for the points of one track.
}
\description{
Read DTI tracking per-coord data from file in MRtrix 'TSF' format.
}
\examples{
\dontrun{
tsff <- "~/simple.tsf"
tsf <- read.dti.tsf(tsff)
}

}
