% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_label.R
\name{read.fs.annot.gii}
\alias{read.fs.annot.gii}
\title{Read an annotation or label in GIFTI format.}
\usage{
read.fs.annot.gii(
  filepath,
  element_index = 1L,
  data_only = FALSE,
  rgb_column_names = c("Red", "Green", "Blue", "Alpha"),
  key_column_name = "Key"
)
}
\arguments{
\item{filepath}{string. Full path to the input label file.}

\item{element_index}{positive integer, the index of the dataarray to return. Ignored unless the file contains several dataarrays.}

\item{data_only}{logical, whether to ignore the colortable and region names, and only return the raw vector that contains one integer label per vertex.}

\item{rgb_column_names}{vector of exactly 4 character strings, order is important. The column names for the red, green, blue and alpha channels in the lable table. If a column does not exist, pass NA. If you do not know the column names, just call the function, it will print them. See 'data_only' if you do not care.}

\item{key_column_name}{character string, the column name for the key column in the lable table. This is the column that holds the label value from the raw vector (see 'data_only') that links a label value to a row in the label table. Without it, one cannot recostruct the region name and color of an entry. Passing NA has the same effect as setting 'data_only' to TRUE.}
}
\description{
Read an annotation or label in GIFTI format.
}
