% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_annot.R
\name{read.fs.annot}
\alias{read.fs.annot}
\title{Read file in FreeSurfer annotation format}
\usage{
read.fs.annot(filepath)
}
\arguments{
\item{filepath, }{string. Full path to the input curv file.}
}
\value{
named list, enties are:
}
\description{
Read a data annotation file in FreeSurfer format. Such a file assigns a label and a color to each vertex of a brain surface. The assignment of labels to vertices is based on at atlas or brain parcellation file. Typically the atlas is available for some standard template subject, and the labels are assigned to another subject by registering it to the template.
   For a subject (MRI image pre-processed with FreeSurfer) named 'bert', an example file would be 'bert/label/lh.aparc.annot', which contains the annotation based on the Desikan-Killiany Atlas for the left hemisphere of bert.
}
