% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.fs.curv}
\alias{read.fs.curv}
\title{Read file in FreeSurfer curv format}
\usage{
read.fs.curv(filepath)
}
\arguments{
\item{filepath, }{string. Full path to the input curv file.}
}
\value{
data, vector of floats. The brain morphometry data, one value per vertex.
}
\description{
Read vertex-wise brain surface data from a file in FreeSurfer binary 'curv' format.
   For a subject (MRI image pre-processed with FreeSurfer) named 'bert', an example file would be 'bert/surf/lh.thickness', which contains n values. Each value represents the cortical thickness at the respective vertex in the brain surface mesh of bert.
}
\examples{
    curvfile = system.file("extdata", "lh.thickness",
                            package = "freesurferformats", mustWork = TRUE);
    ct = read.fs.curv(curvfile);
    cat(sprintf("Read data for \%d vertices. Values: min=\%f, mean=\%f, max=\%f.\\n",
                            length(ct), min(ct), mean(ct), max(ct)));

}
