% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_label.R
\name{read.fs.label.gii}
\alias{read.fs.label.gii}
\title{Read annotation from GIFTI label file}
\usage{
read.fs.label.gii(filepath, label_value, element_index = 1L)
}
\arguments{
\item{filepath}{string. Full path to the input label file.}

\item{label_value}{the label value of interest to extract from the annotation.}

\item{element_index}{positive integer, the index of the dataarray to return. Ignored unless the file contains several dataarrays.}
}
\value{
an annotation, see \code{\link[freesurferformats]{read.fs.annot}}.
}
\description{
Read annotation from GIFTI label file. If all you want from the file is a single label, see \code{\link[freesurferformats]{read.fs.label.gii}}.
}
\seealso{
Other label functions: 
\code{\link{read.fs.label}()},
\code{\link{write.fs.label}()}
}
\concept{label functions}
