% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.fs.morph}
\alias{read.fs.morph}
\title{Read morphometry data file in any FreeSurfer format}
\usage{
read.fs.morph(filepath)
}
\arguments{
\item{filepath, }{string. Full path to the input file. The suffix determines the expected format as follows: ".mgz" and ".mgh" will be read with the read.fs.mgh function, all other file extensions will be read with the read.fs.curv function.}
}
\value{
data, vector of floats. The brain morphometry data, one value per vertex.
}
\description{
Read vertex-wise brain surface data from a file. The file can be in any of the supported formats, and the format will be determined from the file extension.
}
