% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.fs.morph.txt}
\alias{read.fs.morph.txt}
\title{Read morphometry data from plain text file}
\usage{
read.fs.morph.txt(filepath)
}
\arguments{
\item{filepath}{path to a file in plain text format. Such a file contains, on each line, a single float value. This very simply and limited *format* is used by the LGI tool by Lyu et al., and easy to generate in shell scripts.}
}
\value{
numeric vector, the curv data
}
\description{
Read morphometry data from plain text file
}
