% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_patch.R
\name{read.fs.patch.asc}
\alias{read.fs.patch.asc}
\title{Read FreeSurfer ASCII format patch.}
\usage{
read.fs.patch.asc(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input patch file in ASCII surface format.}
}
\value{
named list. The list has the following named entries: "vertices": nx3 double matrix, where n is the number of vertices. Each row contains the x,y,z coordinates of a single vertex. "faces": nx3 integer matrix. Each row contains the vertex indices of the 3 vertices defining the face. WARNING: The indices are returned starting with index 1 (as used in GNU R). Keep in mind that you may need to adjust the index to compare with data from other software that uses zero-based indexing.
}
\description{
An ASCII format patch is a part of a brain surface mesh, and is a mesh itself. It consists of vertices and faces. The ASCII patch format is very similar to the ASCII surface format. **Note:** The contents of ASCII and binary patch format files is different.
}
\seealso{
Other patch functions: 
\code{\link{read.fs.patch}()}
}
\concept{patch functions}
