% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_patch.R
\name{read.fs.patch}
\alias{read.fs.patch}
\title{Read FreeSurfer binary patch file.}
\usage{
read.fs.patch(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input patch file. An example file is `FREESURFER_HOME/subjects/fsaverage/surf/lh.cortex.patch.3d`.}
}
\value{
named list with entry: "data": numerical *n*x7 matrix. The columns are named, and appear in the following order: 'vert_index1': the one-based (R-style) vertex index. 'x', 'y', 'z': float vertex coordinates. 'is_border': integer, 1 if the vertex lies on the patch border, 0 otherwise (treat as logical). 'raw_vtx': integer, the raw vtx value encoding index and border. 'vert_index0': the zero-based (C-style) vertex index.
}
\description{
A binary format patch contains vertices only, without connection (face) information. **Note:** The contents of ASCII and binary patch format files is different.
}
\seealso{
Other patch functions: 
\code{\link{read.fs.patch.asc}()}
}
\concept{patch functions}
