% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.byu}
\alias{read.fs.surface.byu}
\title{Read mesh in BYU format.}
\usage{
read.fs.surface.byu(filepath, part = 1L)
}
\arguments{
\item{filepath}{full path of the file in BYU format.}

\item{part}{positive integer, the index of the mesh that should be loaded from the file. Only relevant if the file contains more than one mesh.}
}
\value{
an `fs.surface` instance, aka a mesh
}
\description{
The BYU format is an old ASCII mesh format that is based on fixed character positions in lines (as opposed to whitespace-separated elements). I consider it a bit counter-intuitive.
}
\references{
\url{http://www.eg-models.de/formats/Format_Byu.html}
}
