% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.gii}
\alias{read.fs.surface.gii}
\title{Read GIFTI format mesh as surface.}
\usage{
read.fs.surface.gii(filepath)
}
\arguments{
\item{filepath}{string. Full path to the input surface file in GIFTI format.}
}
\value{
named list. The list has the following named entries: "vertices": nx3 double matrix, where n is the number of vertices. Each row contains the x,y,z coordinates of a single vertex. "faces": nx3 integer matrix. Each row contains the vertex indices of the 3 vertices defining the face. WARNING: The indices are returned starting with index 1 (as used in GNU R). Keep in mind that you need to adjust the index (by substracting 1) to compare with data from other software.
}
\description{
Read GIFTI format mesh as surface.
}
\seealso{
Other mesh functions: 
\code{\link{read.fs.surface.asc}()},
\code{\link{read.fs.surface.ply}()},
\code{\link{read.fs.surface.vtk}()},
\code{\link{read.fs.surface}()},
\code{\link{read_nisurfacefile}()},
\code{\link{read_nisurface}()},
\code{\link{write.fs.surface.asc}()},
\code{\link{write.fs.surface.gii}()},
\code{\link{write.fs.surface.mz3}()},
\code{\link{write.fs.surface.vtk}()},
\code{\link{write.fs.surface}()}

Other gifti readers: 
\code{\link{read.fs.annot.gii}()},
\code{\link{read.fs.label.gii}()},
\code{\link{read.fs.morph.gii}()}
}
\concept{gifti readers}
\concept{mesh functions}
