% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.mz3}
\alias{read.fs.surface.mz3}
\title{Read surface mesh in mz3 format, used by Surf-Ice.}
\usage{
read.fs.surface.mz3(filepath)
}
\arguments{
\item{filepath}{full path to surface mesh file in mz3 format.}
}
\description{
Read surface mesh in mz3 format, used by Surf-Ice.
}
\references{
See \url{https://github.com/neurolabusc/surf-ice} for details on the format.
}
