% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read.fs.surface.mz3}
\alias{read.fs.surface.mz3}
\title{Read surface mesh in mz3 format, used by Surf-Ice.}
\usage{
read.fs.surface.mz3(filepath)
}
\arguments{
\item{filepath}{full path to surface mesh file in mz3 format.}
}
\value{
an `fs.surface` instance. If the mz3 file contained RGBA per-vertex colors or scalar per-vertex data, these are available in the 'metadata' property.
}
\description{
The mz3 format is a binary file format that can store a mesh (vertices and faces), and optionally per-vertex colors or scalars.
}
\references{
See https://github.com/neurolabusc/surf-ice for details on the format.
}
