% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_curv.R
\name{read.smp.brainvoyager}
\alias{read.smp.brainvoyager}
\title{Read Brainvoyager statistical surface results from SMP file.}
\usage{
read.smp.brainvoyager(filepath)
}
\arguments{
\item{filepath}{character string, path to file in Brainvoyager SMP file format}
}
\value{
named list of file contents
}
\description{
Read Brainvoyager statistical surface results from SMP file.
}
\note{
Currently only SMP file version 3 is supported.
}
\references{
see \url{https://support.brainvoyager.com/brainvoyager/automation-development/84-file-formats/40-the-format-of-smp-files} for the spec
}
