% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_surface.R
\name{read_nisurfacefile.fsascii}
\alias{read_nisurfacefile.fsascii}
\title{Read a FreeSurfer ASCII surface file.}
\usage{
\method{read_nisurfacefile}{fsascii}(filepath, ...)
}
\arguments{
\item{filepath}{character string, the full path to the input surface file.}

\item{...}{parameters passed to \code{\link[freesurferformats]{read.fs.surface.asc}}.}
}
\value{
an instance of `fs.surface`, read from the file. See \code{\link[freesurferformats]{read.fs.surface}} for details. If none of the reader methods succeed, an error is raised.
}
\description{
Read a FreeSurfer ASCII surface file.
}
