% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{sm0to1}
\alias{sm0to1}
\title{Adapt spatial transformation matrix for 1-based indices.}
\usage{
sm0to1(tf_matrix)
}
\arguments{
\item{tf_matrix}{4x4 numerical matrix, the input spatial transformation matrix, suitable for 0-based indices. Typically this is a vox2ras matrix obtained from functions like \code{\link{mghheader.vox2ras}}.}
}
\value{
4x4 numerical matrix, adapted spatial transformation matrix, suitable for 1-based indices.
}
\description{
Adapt spatial transformation matrix for 1-based indices.
}
