% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{sm0to1}
\alias{sm0to1}
\title{Adapt spatial transformation matrix for 1-based indices.}
\usage{
sm0to1(tf_matrix)
}
\arguments{
\item{tf_matrix}{4x4 numerical matrix, the input spatial transformation matrix, suitable for 0-based indices. Typically this is a vox2ras matrix obtained from functions like \code{\link{mghheader.vox2ras}}.}
}
\value{
4x4 numerical matrix, adapted spatial transformation matrix, suitable for 1-based indices.
}
\description{
Adapt spatial transformation matrix for 1-based indices.
}
\seealso{
\code{\link{sm1to0}} for the inverse operation

Other header coordinate space: 
\code{\link{mghheader.is.ras.valid}()},
\code{\link{mghheader.ras2vox.tkreg}()},
\code{\link{mghheader.ras2vox}()},
\code{\link{mghheader.scanner2tkreg}()},
\code{\link{mghheader.tkreg2scanner}()},
\code{\link{mghheader.vox2ras.tkreg}()},
\code{\link{mghheader.vox2ras}()},
\code{\link{read.fs.transform.dat}()},
\code{\link{read.fs.transform.xfm}()},
\code{\link{read.fs.transform}()},
\code{\link{sm1to0}()}
}
\concept{header coordinate space}
