% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{sm1to0}
\alias{sm1to0}
\title{Adapt spatial transformation matrix for 0-based indices.}
\usage{
sm1to0(tf_matrix)
}
\arguments{
\item{tf_matrix}{4x4 numerical matrix, the input spatial transformation matrix, suitable for 1-based indices.}
}
\value{
4x4 numerical matrix, adapted spatial transformation matrix, suitable for 0-based indices.
}
\description{
Adapt spatial transformation matrix for 0-based indices.
}
\seealso{
\code{\link{sm0to1}} for the inverse operation

Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.dat]{read.fs.transform.dat()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform.to.voxel]{transform.to.voxel()}},
\code{\link[=transform.to.world]{transform.to.world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.dat]{write.fs.transform.dat()}},
\code{\link[=write.fs.transform.fslmat]{write.fs.transform.fslmat()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}},
\code{\link[=write.fs.transform.xfm]{write.fs.transform.xfm()}}
}
\concept{header coordinate space}
