% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{sm1to0}
\alias{sm1to0}
\title{Adapt spatial transformation matrix for 0-based indices.}
\usage{
sm1to0(tf_matrix)
}
\arguments{
\item{tf_matrix}{4x4 numerical matrix, the input spatial transformation matrix, suitable for 1-based indices.}
}
\value{
4x4 numerical matrix, adapted spatial transformation matrix, suitable for 0-based indices.
}
\description{
Adapt spatial transformation matrix for 0-based indices.
}
