% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mghheader.R
\name{talairachras.to.ras}
\alias{talairachras.to.ras}
\title{Compute MNI talairach coordinates from RAS coords.}
\usage{
talairachras.to.ras(tal_ras_coords, talairach)
}
\arguments{
\item{tal_ras_coords}{coordinate matrix in Talairach RAS space}

\item{talairach}{the 4x4 numerical talairach matrix, or a character string which will be interpreted as the path to an xfm file containing the matrix (typically `$SUBJECTS_DIR/$subject/mri/transforms/talairach.xfm`).}
}
\value{
the Talairach RAS coordinates for the given RAS coordinates. They are based on a linear transform.
}
\description{
Compute MNI talairach coordinates from RAS coords.
}
\note{
You can use this to compute the Talairach coordinate of a voxel, based on its RAS coordinate.
}
\references{
see \url{https://en.wikipedia.org/wiki/Talairach_coordinates}
}
