% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_dist.R
\name{vertex.euclid.dist}
\alias{vertex.euclid.dist}
\title{Compute Euclidian distance between two vertices v1 and v2.}
\usage{
vertex.euclid.dist(surface, v1, v2)
}
\arguments{
\item{surface}{an fs.surface instance}

\item{x1}{positive integer, vertex index of v1}

\item{x2}{positive integer, vertex index of v2}
}
\value{
the Euclidian distance between v1 and v2.
}
\description{
Compute Euclidian distance between two vertices v1 and v2.
}
