% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_dist.R
\name{vertex.euclid.dist}
\alias{vertex.euclid.dist}
\title{Compute Euclidean distance between two vertices v1 and v2.}
\usage{
vertex.euclid.dist(surface, v1, v2)
}
\arguments{
\item{surface}{an fs.surface instance}

\item{v1}{positive integer, vertex index of v1}

\item{v2}{positive integer, vertex index of v2}
}
\value{
double, the Euclidean distance between v1 and v2
}
\description{
Compute Euclidean distance between two vertices v1 and v2.
}
\seealso{
Other Euclidean distance util functions: 
\code{\link{closest.vert.to.point}()},
\code{\link{vertexdists.to.point}()}
}
\concept{Euclidean distance util functions}
