% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{vis.quakemodel}
\alias{vis.quakemodel}
\title{Visualize Quake II alias model.}
\usage{
vis.quakemodel(md2, texture_file = NULL, frame_idx = 1L)
}
\arguments{
\item{md2}{an MD2 model instance.}

\item{texture_file}{character string, path to model skin.}

\item{frame_idx}{integer, which frame to use for vertex positions. A model contains many vertex positions if it includes model animation data.}
}
\description{
Visualize Quake II alias model.
}
