% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_surface.R
\name{write.fs.surface.off.ply2}
\alias{write.fs.surface.off.ply2}
\title{Write mesh to file in Object File Format (.off) or PLY2 format.}
\usage{
write.fs.surface.off.ply2(filepath, vertex_coords, faces, format)
}
\arguments{
\item{filepath}{string. Full path to the output surface file, should end with '.off', but that is not enforced.}

\item{vertex_coords}{n x 3 matrix of doubles. Each row defined the x,y,z coords for a vertex.}

\item{faces}{n x 3 matrix of integers. Each row defined the 3 vertex indices that make up the face. WARNING: Vertex indices should be given in R-style, i.e., the index of the first vertex is 1. However, they will be written in FreeSurfer style, i.e., all indices will have 1 substracted, so that the index of the first vertex will be zero.}

\item{format}{character string, the format to write. One of 'ply2' or 'off'.}
}
\value{
string the format that was written. One of "tris" or "quads". Currently only triangular meshes are supported, so always 'tris'.
}
\description{
The two formats are very similar, they only differ in the header lines. This function can write both.
}
\note{
Do not confuse the OFF format (.off) with the Wavefront object file format (.obj), they are not identical.
}
\examples{
\dontrun{
# Read a surface from a file:
surface_file <- system.file("extdata", "lh.tinysurface",
  package = "freesurferformats", mustWork = TRUE
)
mesh <- read.fs.surface(surface_file)
# Now save it:
write.fs.surface.off(tempfile(fileext = ".off"), mesh$vertices, mesh$faces)
}

}
\seealso{
Other mesh export functions:
\code{\link[=write.fs.surface]{write.fs.surface()}},
\code{\link[=write.fs.surface.obj]{write.fs.surface.obj()}},
\code{\link[=write.fs.surface.off]{write.fs.surface.off()}},
\code{\link[=write.fs.surface.ply]{write.fs.surface.ply()}},
\code{\link[=write.fs.surface.ply2]{write.fs.surface.ply2()}},
\code{\link[=write.fs.surface.stl]{write.fs.surface.stl()}},
\code{\link[=write.fs.surface.vtk]{write.fs.surface.vtk()}}
}
\concept{mesh export functions}
\keyword{internal}
