% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_transform.R
\name{write.fs.transform.dat}
\alias{write.fs.transform.dat}
\title{Write a tkregister dat file.}
\usage{
write.fs.transform.dat(
  tf,
  filepath,
  subject = NULL,
  in_plane_resolution = NULL
)
}
\arguments{
\item{tf}{an \code{fs.transform} instance whose matrix maps voxel coordinates to RAS coordinates.}

\item{filepath}{character string, the full path of the file to write.}

\item{subject}{\code{NULL} or character string, the subject identifier to store in the first line of the file. This
is metadata for the tools that read the file and does not influence the transformation. If \code{NULL} and \code{tf}
has a \code{subject} field (as read by \code{\link{read.fs.transform.dat}}), that value is used.}

\item{in_plane_resolution}{\code{NULL} or numerical vector of length 2, the in-plane and between-plane resolution
of the movable volume in millimeters, stored in the second and third line of the file. These are metadata
that describe the acquisition and do not influence the transformation. If \code{NULL} and \code{tf} has the
corresponding fields, they are used, otherwise they are derived from the geometry of the source volume, and
set to 1 if that is not available either.}
}
\value{
the \code{fs.transform} instance \code{tf}, invisibly.
}
\description{
Write a transformation in the FreeSurfer tkregister format (\code{register.dat}), the format that
\code{mri_vol2vol --reg}, \code{tkregister2} and \code{bbregister} use. Such a matrix maps the voxel coordinates of the
movable volume (the source) to RAS coordinates in the tkregister frame of the target volume, see
\code{\link{mghheader.vox2ras.tkreg}}, so a transformation can only be written if this is what it maps.
}
\examples{
dat_file <- system.file("extdata", "register.dat", package = "freesurferformats", mustWork = TRUE)
tf <- read.fs.transform(dat_file)
out_file <- tempfile(fileext = ".dat")
write.fs.transform.dat(tf, out_file)
max(abs(read.fs.transform(out_file)$matrix - tf$matrix)) # 0
unlink(out_file)

}
\seealso{
Other header coordinate space:
\code{\link[=invert.fs.transform]{invert.fs.transform()}},
\code{\link[=is.fs.transform]{is.fs.transform()}},
\code{\link[=mghheader.is.ras.valid]{mghheader.is.ras.valid()}},
\code{\link[=mghheader.ras2vox]{mghheader.ras2vox()}},
\code{\link[=mghheader.ras2vox.tkreg]{mghheader.ras2vox.tkreg()}},
\code{\link[=mghheader.scanner2tkreg]{mghheader.scanner2tkreg()}},
\code{\link[=mghheader.tkreg2scanner]{mghheader.tkreg2scanner()}},
\code{\link[=mghheader.vox2ras]{mghheader.vox2ras()}},
\code{\link[=mghheader.vox2ras.tkreg]{mghheader.vox2ras.tkreg()}},
\code{\link[=print.fs.transform]{print.fs.transform()}},
\code{\link[=read.fs.transform]{read.fs.transform()}},
\code{\link[=read.fs.transform.dat]{read.fs.transform.dat()}},
\code{\link[=read.fs.transform.fslmat]{read.fs.transform.fslmat()}},
\code{\link[=read.fs.transform.itk]{read.fs.transform.itk()}},
\code{\link[=read.fs.transform.lta]{read.fs.transform.lta()}},
\code{\link[=read.fs.transform.xfm]{read.fs.transform.xfm()}},
\code{\link[=sm0to1]{sm0to1()}},
\code{\link[=sm1to0]{sm1to0()}},
\code{\link[=summary.fs.transform]{summary.fs.transform()}},
\code{\link[=transform.to.lps]{transform.to.lps()}},
\code{\link[=transform.to.ras]{transform.to.ras()}},
\code{\link[=transform.to.voxel]{transform.to.voxel()}},
\code{\link[=transform.to.world]{transform.to.world()}},
\code{\link[=write.fs.transform]{write.fs.transform()}},
\code{\link[=write.fs.transform.fslmat]{write.fs.transform.fslmat()}},
\code{\link[=write.fs.transform.itk]{write.fs.transform.itk()}},
\code{\link[=write.fs.transform.lta]{write.fs.transform.lta()}},
\code{\link[=write.fs.transform.xfm]{write.fs.transform.xfm()}}
}
\concept{header coordinate space}
