% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti1.R
\name{write.nifti1}
\alias{write.nifti1}
\title{Write header and data to a file in NIFTI v1 format.}
\usage{
write.nifti1(niiheader, niidata)
}
\arguments{
\item{niiheader}{a NIFTI v1 header that is suitable for the passed 'niidata'.}

\item{niidata}{array of numeric or integer data, with up to 7 dimensions. Will be written to the file with the datatype and bitpix specified in the 'niiheader' argument.}
}
\description{
Write header and data to a file in NIFTI v1 format.
}
